==> rmd/03_northern_lats_EDA_cc_env_figs.Rmd <==
---
title: "3 | Turtle-Env Figs"
subtitle: "Project STRETCH: Northern Latitudes - Exploratory Analyses"
author: "Dana K Briscoe"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: none
    df_print: paged
    highlight: tango
    number_sections: no
    theme: flatly
    toc: no
    toc_float: true
    toc_collapsed: true
  pdf_document:
    toc: yes
    toc_depth: '2'
---

```{=html}
<style>
.main-container {
    max-width: 1180px;
    margin-left: 10;
    margin-right: 10;
}
</style>
```
```{r 03-setup, include=FALSE, warning=FALSE}
knitr::opts_chunk$set(
    echo = FALSE,
    eval = TRUE,
    message = FALSE,
    warning = FALSE,
    dpi = 300,
    fig.align = "center"
    )

# See Options: knitr::opts_chunk$get()
```

```{r 03-load-libraries, message=FALSE}
# 01
library(tidyverse)
library(data.table)
library(ggplot2)
library(plotly)
library(leaflet)
library(mapview)
library(leafsync)
library(leafgl)
library(sf)
library(kableExtra)
library(knitr)
# 02
library(lubridate)
library(scales)
# 03 
library(RColorBrewer)
library(fields)
library(colorRamps)
require(latticeExtra)
require(zoo)
suppressMessages(library(raster))
library(glue)
```

```{r 03-source-helper-funcs}
source('../code/00_northern_lats_helper_functions.R')
```

```{r 03-load-cohort-1}
source('../code/01_prep_turtle_data.R')

```


<!-- ## Load Data -->
```{r 03-rename-dfs-for-consistency}
raw_df <- raw_data_cohort_1_w_names
daily_df <- daily_avg_data_cohort_1

# make historic df consistent with cohort df(s)
hist_df <-  historic_tags %>% wrangleHistDF()

```


```{r 03-set-nc-path}
nc_path = "/Users/briscoedk/dbriscoe@stanford.edu - Google Drive/My Drive/ncdf/npac"

```


```{r 03-load-interim-data}
f <- "../data/interim/03_northern_lats_EDA_interim_data.RData"
if (file.exists(f)){
    load(f)
}

```


```{r 03-set-spatial-extent}
# bbox <- c(make360(-170), make360(-110), 20, 50)
bbox <- c(-170, -110, 20, 50)
# bbox <- c(-170, -110, 20, 50)
e <- extent(bbox[1], bbox[2], bbox[3], bbox[4])

```

```{r 03-set-ras-params}
params <- tibble(
    # eov = "ssta",
    timestep = "month",
    min_dt = '2023-07-16',
    max_dt = '2023-09-16')

```

```{r 03-load ssta-monthly-ncdf}
params$eov <- "ssta"

ssta_ncs <- list.files(nc_path, pattern = str_c(params$eov, "_"), full.names=T) %>%
    .[grep(params$timestep, .)]


ssta_fdates <- ssta_ncs %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 6)) %>%
    substr(., start=1, stop=10)

      
```

```{r 03-load chl-monthly-ncdf}
params$eov <- "chla"

chla_ncs <- list.files(nc_path, pattern = str_c(params$eov, "_"), full.names=T) %>%
    .[grep(params$timestep %>% str_to_title(), .)]


chla_fdates <- chla_ncs %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 3)) %>%
    substr(., start=1, stop=10)


```

```{r 03-load sst-monthly-ncdf}
params$eov <- "sst"

sst_ncs <- list.files(nc_path, pattern = str_c(params$eov, "_"), full.names=T) %>%
    .[grep(params$timestep, .)]


sst_fdates <- sst_ncs %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 6)) %>%
    substr(., start=1, stop=10)

      
```

```{r 03-stack-rasters-2023-only, warning=FALSE,message=FALSE, results='hide'}

if(!exists("ssta_ras_stack")){

dt_idx <- getDateRange(params$min_dt, params$max_dt, unit = params$timestep)

## ssta
ssta_ras_stack <- ssta_ncs[ssta_fdates >= params$min_dt & ssta_fdates <= params$max_dt] %>%
    raster::stack(.) %>%
    rotate(.) %>% # convert into -`180 to 180
    crop(., e) %>%
    prepRasterZ(., dt_idx) 

}

if(!exists("chla_ras_stack")){
## chla
chla_ras_stack <- chla_ncs[chla_fdates >= params$min_dt & chla_fdates <= params$max_dt] %>%
    raster::stack(., varname="chlor_a") %>%
    crop(., e) %>% 
    # shift(rotate(shift(., 180)), 180) %>% 
    prepRasterZ(., dt_idx) %>%
    invisible()
}

```

```{r 03-stack-rasters-1997-2023, warning=FALSE,message=FALSE, results='hide'}

if(!exists("sst_ras_stack")){
# fyi, since getting full nc range, don't need to subset dates

sst_dt_idx <- sst_fdates #getDateRange(params$min_dt, params$max_dt, unit = params$timestep)

## ssta
sst_ras_stack <- sst_ncs %>% #. [sst_fdates >= params$min_dt & sst_fdates <= params$max_dt] %>%
    raster::stack(.) %>%
    rotate(.) %>% # convert into -`180 to 180
    crop(., e) %>%
    prepRasterZ(., sst_dt_idx) 
}

```


<!-- ```{r intermediate-step-calc-sept2023-from-dailys} -->

<!-- if(!exists("daily_sst_ras_stack")){ -->

<!-- nc_path <- "/Users/briscoedk/dbriscoe@stanford.edu - Google Drive/My Drive/ncdf/deploy_reports" -->
<!-- daily_sst_ncs <- list.files(nc_path, pattern =  "sst_", full.names=T) -->
<!-- daily_sst_fdates <- daily_sst_ncs %>%  -->
<!--         str_split(., "_") %>% -->
<!--         purrr::map_chr(~ pluck(., 5)) %>% -->
<!--         substr(., start=1, stop=10) -->

<!-- daily_sst_dt_idx <- getDateRange('2023-09-01', '2023-09-19', unit = 'daily') -->

<!-- ## ssta -->
<!-- daily_sst_ras_stack <- daily_sst_ncs[daily_sst_fdates >= min(daily_sst_dt_idx) & daily_sst_fdates <= max(daily_sst_dt_idx)] %>% -->
<!--     raster::stack(.) %>% -->
<!--     rotate(.) %>% # convert into -`180 to 180 -->
<!--     crop(., e) %>% -->
<!--     prepRasterZ(., daily_sst_dt_idx)  -->

<!-- # calc temp ras sept 2023 -->
<!-- ras_sept_2023 <- calc(daily_sst_ras_stack, fun = mean, na.rm = T) -->

<!-- names(ras_sept_2023) <- c('X2023-09-16') -->

<!-- } -->
<!-- ``` -->



```{r 03-calc-ras-mean-layer}
## ssta
ssta_ras_mean <- ssta_ras_stack %>%
     calc(., mean, na.rm=TRUE)

names(ssta_ras_mean) = as.character(dt_idx |> (\(x) { tail(x, n=1)})() )# assign dummy date for mean ras. use last date

## chla
chla_ras_mean <- chla_ras_stack %>%
     calc(., mean, na.rm=TRUE)

names(chla_ras_mean) = as.character(dt_idx |> (\(x) { tail(x, n=1)})() )# assign dummy date for mean ras. use last date


sst_ras_mean <- sst_ras_stack %>%
    calc(., mean, na.rm=TRUE)

names(sst_ras_mean) = as.character(dt_idx |> (\(x) { tail(x, n=1)})() )# assign dummy date for mean ras. use last date

```

```{r 03-convert-ras-to-df}
# all-time mean df
ssta_ras_mean_df <- ssta_ras_mean %>% 
    ras2df()

# monthly df
ssta_ras_monthly_df <- ssta_ras_stack %>% ras2df()

## chla
# all-time mean df
chla_ras_mean_df <- chla_ras_mean %>% 
    ras2df()

# monthly df
chla_ras_monthly_df <- chla_ras_stack %>% ras2df()


# sst_ras_mean_df <- sst_ras_stack %>% ras2df() ## dkb note: memory exceeded
    
```

<!-- ## Set Plot Params -->

```{r 03-set-plot-params}
# mapdata <- map_data('world', wrap=c(-25,335), ylim=c(-55,75)) %>%
#     filter(long >= 120 & long <= 270 & lat >= 15 & lat <=80) 
mapdata <- map_data('world')

grid_theme <- theme_bw() + theme(panel.ontop=TRUE, panel.background=element_blank())


ssta_pal <- rev(c(rev(brewer.pal(9, "YlOrRd")),"white", brewer.pal(9, "Blues")))
# chla_pal <- brewer.pal(9, "Greens")
chla_pal <- oce::oceColorsChlorophyll(20)

wc_pal   <- rainbow(25)

smooth_rainbow <- khroma::colour("smooth rainbow")

sst_cpal <- c(smooth_rainbow(length(seq(floor(8), ceiling(32), 1)), range = c(0, 0.9)), "#9e2a2b", "firebrick4", "#540b0e")

```

```{r 03-make-baseplot}
baseplot <- ggplot() + 
        # add coast 
    geom_polygon(data = mapdata, aes(x=long, y = lat, group = group), color = "black", fill = "black") +
   
    # # theme_minimal() +
    grid_theme +
    #   theme(panel.background = element_rect(fill = NA),
    #   panel.grid.major = element_line(colour = 'snow', linetype = "solid", size = 0.5),panel.ontop = TRUE) +
    
    coord_sf(xlim = c(bbox[1], bbox[2]), ylim = c(bbox[3], bbox[4]), 
             expand = FALSE, crs = st_crs(4326)) +
labs(x = "Longitude", y = "Latitude")

```

```{r 03-make-gglayers}
add_baseplot <- function(bbox) {
    list(
    # add land
    geom_polygon(data = mapdata, aes(x=long, y = lat, group = group), color = "black", fill = "black"),
    # sest cbar height
    guides(fill = guide_colourbar(barheight =12, ticks = TRUE,
                  frame.colour = "black", ticks.colour = "black")),
    # set map extent
    coord_sf(xlim = c(bbox[1], bbox[2]), ylim = c(bbox[3], bbox[4]), 
             expand = FALSE, crs = st_crs(4326)), 
    labs(x = "Longitude", y = "Latitude")
    )
}

update_map_extent <- function(xrange, yrange){
    list(
        coord_sf(xlim = c(xrange[1], xrange[2]), ylim = c(yrange[1], yrange[2]), 
        expand = FALSE, crs = st_crs(4326))
    )
}

add_turtle_pts <- function(df, col="#66CCC0", size=3){
    list(
        geom_point(data=df, aes(x = lon, y = lat, colour = col), size = size)
    )
}

# add_turtle_pts <- function(df, size=3){
#     list(
#         geom_point(data=df, aes(x = lon, y = lat, colour = as.factor(group)), size = size)
#     )
# }

add_track_spLines <- function(df, cpal, linewidth) {
    list(
    # add cohort 1 tracks
    geom_sf(data = df %>% makeSpatialLines(lon360=FALSE), 
        aes(colour = as.factor(id)),linewidth = linewidth, alpha = 0.9, show.legend=FALSE),
        scale_colour_manual(values = cpal)
    )
}

add_tcms_box <- function(){
    list(
    geom_rect(aes(xmin=make180(225),xmax=make180(243),ymin=25,ymax=35),
          colour="honeydew3",alpha=0, size=0.45)
    )
}


add_contour_lines <- function(df, contours, col='black'){
    list(
        geom_contour(data=df, aes(x=lon, y=lat, z = val), 
        colour = col, linewidth = 1.25, breaks = c(contours), show.legend = F) 
    )
}

add_contour_lines_dashed <- function(df, contours, col='black'){
    list(
        geom_contour(data=df, aes(x=lon, y=lat, z = val), linetype = 'dashed', 
        colour = col, linewidth = 1.25, breaks = c(contours), show.legend = F) 
    )
}

resize_cbar <- function(x){
    list(
        guides(fill = guide_colourbar(barheight =x, ticks = TRUE,
        frame.colour = "black", ticks.colour = "black"))
    )
}

```

<!-- ------------------------------------------------------------------------ -->

```{r 03-plot-ssta-mean-no-cc-tracks}

p_ssta_mean_no_cc <- ggplot() +
    # add ssta layer
    geom_raster(data = ssta_ras_mean_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    scale_fill_gradientn(colours = ssta_pal[5:length(ssta_pal)],
    # scale_colour_gradient2(high = "red", mid = "white", low = "blue", midpoint = 1, 
                         breaks=seq(-2,4,1),
                         limits = c(-2,4), 
                         na.value="transparent",
                         name = "SSTA (°C) \n") +
    # # add cohort 1 tracks
    # add_track_spLines(df=daily_df %>% filter(date <='2023-09-30'), cpal=wc_pal, linewidth=1) +

    # add baselayers
    add_baseplot(bbox) 
 
# show plot with tcms box
p_ssta_mean_no_cc_or_tcms <- p_ssta_mean_no_cc #+ add_tcms_box()
# p_ssta_mean_no_cc_with_tcms <- p_ssta_mean_no_cc + add_tcms_box()


# save
ggsave(filename = str_c('plot_ssta_mean_no_cc_or_tcms_w_contour.png'), plot = p_ssta_mean_no_cc_or_tcms + resize_cbar(24) +
           labs(caption = ''), 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

# ggsave(filename = str_c('plot_ssta_mean_no_cc_with_tcms_w_contour.png'), plot = p_ssta_mean_no_cc_with_tcms + 
#            labs(caption = 'Black line represents the 1.5°C SSTA contour'), 
#        # height=9,
#        width=13, bg='white',
#        path = '../figs/'
#        )

```

<!-- ## Part 1: SSTA Map with Tracks -->

```{r 03-plot-ssta-mean-cc-tracks}

p_ssta_mean <- ggplot() +
    # add ssta layer
    geom_raster(data = ssta_ras_mean_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    scale_fill_gradientn(colours = ssta_pal[5:length(ssta_pal)],
    # scale_colour_gradient2(high = "red", mid = "white", low = "blue", midpoint = 1, 
                         breaks=seq(-2,4,1),
                         limits = c(-2,4), 
                         na.value="transparent",
                         name = "SSTA (°C) \n") +
    # add cohort 1 tracks
    add_track_spLines(df=daily_df %>% filter(date <='2023-09-30'), cpal=wc_pal, linewidth=1) +

    # add baselayers
    add_baseplot(bbox) 
 
# show plot with tcms box
p_ssta_mean_tcms <- p_ssta_mean + add_tcms_box()
# p_ssta_mean_tcms

```

```{r 03-plot-ssta-cohort-1-with-contour}
p_ssta_mean_tcms_w_contour <- p_ssta_mean_tcms + 
    add_contour_lines(df=ssta_ras_mean_df, contours=c(1.5)) + 
    resize_cbar(24)

# save
ggsave(filename = str_c('plot_cc_ssta_mean_tcms_w_contour.png'), plot = p_ssta_mean_tcms_w_contour + 
           labs(caption = 'Black line represents the 1.5°C SSTA contour'), 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

```

```{r 03-plot-ssta-cohort-1-with-contour-zoomed}
p_ssta_mean_tcms_w_contour_zoom <- p_ssta_mean_tcms_w_contour +
    update_map_extent(xrange=c(-155, -130), yrange = c(35, 50))

```

<!-- ------------------------------------------------------------------------ -->

<!-- ## Part 2: Chla Map with Tracks -->

```{r 03-plot-chla-mean-tcms-cohort1}
    
p_chla_mean <- ggplot() +
    # add ssta layer
    geom_raster(data = chla_ras_mean_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    scale_fill_gradientn(colours = rev(wc_pal[1:(length(wc_pal)-4)]), trans="log",
    # scale_fill_gradientn(colours = chla_pal[1:length(chla_pal)], trans="log",

    # scale_fill_continuous(type = "viridis", trans="log",
                      # breaks=c(0.05, 0.1, 0.2, 0.4, 1, 2, 5, 10, 30),
                          breaks=c(0.02, 0.05, 0.1, 0.2, 0.4, 1, 2, 5, 10, 30),
    
                      # breaks=c(0.02, 0.04, 0.1, 0.2, 0.4, 1, 2, 4, 10),
                         limits = c(0.019,30),
                         na.value="transparent",
                         name = "Chla (mg/m^3) \n") +
    # add cohort 1 tracks
    add_track_spLines(daily_df %>% filter(date <= '2023-09-30'), cpal=wc_pal, linewidth=1) +

    # add baselayers
    add_baseplot(bbox) 
 
# show plot with tcms box
p_chla_mean_tcms <- p_chla_mean + add_tcms_box() +
    resize_cbar(13)
    #+ add_contour_lines(chla_ras_mean_df, contours = 0.20)

```

```{r 03-plot-chla-mean-tcms-cohort1-zoomed}
p_chla_mean_tcms_zoom <- 
p_chla_mean_tcms +
        update_map_extent(xrange=c(-155, -130), yrange = c(35, 50))
```

<!-- ------------------------------------------------------------------------ -->

<!-- ## Part 3: Map TZCF isotherm location [Historic (1997-2012) vs 2023] across eastern NPac -->

```{r 03-pull-yr-mon-from-ras}

names(sst_ras_stack) <- getZ(sst_ras_stack)

yrs <- getZ(sst_ras_stack) |> format("%Y") |> unique()
mons <- getZ(sst_ras_stack) |> format("%b") |> unique()

dte <- getZ(sst_ras_stack)
m <- format(dte, "%b")
y <- format(dte, "%Y")

```


```{r 03-get-sept-ras-cc}

# 1) get historic Sept months: 1997-2012
i_hist_sept <- dte[(m == "Sep" & y <= "2012")] %>% gsub(" ", ".", .)
ras_sept_1997_2012 <- sst_ras_stack[[i_hist_sept]]

# 2) get historic subset Sept months: 2005-2006
i_hist_sept_2005_2006 <- dte[(m == "Sep" & y >= "2005" & y <= "2006")] %>% gsub(" ", ".", .)
ras_sept_2005_2006 <- sst_ras_stack[[i_hist_sept_2005_2006]]


# 3) get cohort 1 Sept month: 2023
i_cohort1_sept <- dte[(m == "Sep" & y == "2023")] %>% gsub(" ", ".", .)
ras_sept_2023 <- sst_ras_stack[[i_cohort1_sept]]

# 4) all time septs (1997-2023)
i_sept <- dte[(m == "Sep")] %>% gsub(" ", ".", .)
ras_sept_1997_2023 <- sst_ras_stack[[i_sept]]


# 5) Longterm avgs (1985/1990 vs 2020-2023)
# 1985-1995
i_hist_sept_1985_1995 <- dte[(m == "Sep" & y >= "1985" & y <= "1995")] %>% gsub(" ", ".", .)
ras_sept_1985_1995 <- sst_ras_stack[[i_hist_sept_1985_1995]]

# 1990-1994
i_hist_sept_1990_1994 <- dte[(m == "Sep" & y >= "1990" & y <= "1994")] %>% gsub(" ", ".", .)
ras_sept_1990_1994 <- sst_ras_stack[[i_hist_sept_1990_1994]]

# 2020-2023
i_hist_sept_2020_2023 <- dte[(m == "Sep" & y >= "2020" & y <= "2023")] %>% gsub(" ", ".", .)
ras_sept_2020_2023 <- sst_ras_stack[[i_hist_sept_2020_2023]]


```

```{r 03-get-march-ras-cc}
# 1) get historic March months: 1997-2012
i_hist_march <- dte[(m == "Mar" & y <= "2012")] %>% gsub(" ", ".", .)
ras_march_1997_2012 <- sst_ras_stack[[i_hist_march]]

# 2) get historic subset March months: 2005-2006
i_hist_march_2005_2006 <- dte[(m == "Mar" & y >= "2005" & y <= "2006")] %>% gsub(" ", ".", .)
ras_march_2005_2006 <- sst_ras_stack[[i_hist_march_2005_2006]]


# 3) get cohort 1 Sept month: 2023
i_march_2023 <- dte[(m == "Mar" & y == "2023")] %>% gsub(" ", ".", .)
ras_march_2023 <- sst_ras_stack[[i_march_2023]]

# 4) all time March (1997-2023)
i_march <- dte[(m == "Mar")] %>% gsub(" ", ".", .)
ras_march_1997_2023 <- sst_ras_stack[[i_march]]

```

```{r 03-get-aug-ras-cc}
# 1) get historic Aug months: 1997-2012
i_hist_aug <- dte[(m == "Aug" & y <= "2012")] %>% gsub(" ", ".", .)
ras_aug_1997_2012 <- sst_ras_stack[[i_hist_aug]]

# 2) get historic subset March months: 2006-2007 --- NOTE!!! NEED TO CHANGE SEPT/MARS from 2005/2006 TO 2006/2007!!!!
i_hist_aug_2006_2007 <- dte[(m == "Aug" & y >= "2006" & y <= "2007")] %>% gsub(" ", ".", .)
ras_aug_2006_2007 <- sst_ras_stack[[i_hist_aug_2006_2007]]

# 3) get cohort 1 Aug month: 2023
i_cohort1_aug <- dte[(m == "Aug" & y == "2023")] %>% gsub(" ", ".", .)
ras_aug_2023 <- sst_ras_stack[[i_cohort1_aug]]

# 4) all time aug (1997-2023)
i_aug <- dte[(m == "Aug")] %>% gsub(" ", ".", .)
ras_aug_1997_2023 <- sst_ras_stack[[i_aug]]


```


```{r 03-calc-sept-ras-means}

# historic septs only
ras_mean_sept_1997_2012 <- calc(ras_sept_1997_2012, fun = mean, na.rm = T)

# historic subset septs only
ras_mean_sept_2005_2006 <- calc(ras_sept_2005_2006, fun = mean, na.rm = T)

# 2023 sept mean only 
# ras_sept_2023 

# # alltime mean (sept 1997 - 2023)  
ras_mean_sept_1997_2023 <- calc(ras_sept_1997_2023, fun = mean, na.rm = T)

```

<!-- ```{r intermediate-step-calc-sept2023-from-dailys} -->

<!-- f <- "../data/interim/03_northern_lats_EDA_interim_ras_sst_data.RData" -->
<!-- if (file.exists(f)){ -->
<!--     load(f) -->
<!-- } else{ -->

<!-- nc_path <- "/Users/briscoedk/dbriscoe@stanford.edu - Google Drive/My Drive/ncdf/deploy_reports" -->
<!-- daily_sst_ncs <- list.files(nc_path, pattern =  "sst_", full.names=T) -->
<!-- daily_sst_fdates <- daily_sst_ncs %>%  -->
<!--         str_split(., "_") %>% -->
<!--         purrr::map_chr(~ pluck(., 5)) %>% -->
<!--         substr(., start=1, stop=10) -->

<!-- daily_sst_dt_idx <- getDateRange('2023-09-01', '2023-09-19', unit = 'daily') -->

<!-- ## ssta -->
<!-- daily_sst_ras_stack <- daily_sst_ncs[daily_sst_fdates >= min(daily_sst_dt_idx) & daily_sst_fdates <= max(daily_sst_dt_idx)] %>% -->
<!--     raster::stack(.) %>% -->
<!--     rotate(.) %>% # convert into -`180 to 180 -->
<!--     crop(., e) %>% -->
<!--     prepRasterZ(., daily_sst_dt_idx)  -->

<!-- # calc temp ras sept 2023 -->
<!-- ras_sept_2023 <- calc(daily_sst_ras_stack, fun = mean, na.rm = T) -->

<!-- names(ras_sept_2023) <- c('X2023-09-16') -->

<!-- ## INTERIM SAVE  -->
<!--     save(daily_sst_ras_stack, ras_sept_2023, file = "03_northern_lats_EDA_interim_ras_sst_data.RData") -->
<!-- } -->

<!-- ``` -->

```{r 03-calc-march-ras-means}

# historic septs only
ras_mean_march_1997_2012 <- calc(ras_march_1997_2012, fun = mean, na.rm = T)

# historic subset septs only
ras_mean_march_2005_2006 <- calc(ras_march_2005_2006, fun = mean, na.rm = T)

# 2023 march mean only 
# ras_march_2023 

# # alltime mean (sept 1997 - 2023)  
ras_mean_march_1997_2023 <- calc(ras_march_1997_2023, fun = mean, na.rm = T)

```

```{r 03-calc-aug-ras-means}

# historic septs only
ras_mean_aug_1997_2012 <- calc(ras_aug_1997_2012, fun = mean, na.rm = T)

# historic subset septs only
ras_mean_aug_2006_2007 <- calc(ras_aug_2006_2007, fun = mean, na.rm = T)

# 2023 sept mean only 
# ras_sept_2023 

# # alltime mean (sept 1997 - 2023)  
ras_mean_aug_1997_2023 <- calc(ras_aug_1997_2023, fun = mean, na.rm = T)

```


```{r 03-ras-sept-means-to-df}

## 1) sept 1997-2012 -- DKB NOTE: Changed from 2013 to 2012, since there are no sept tracks in 2013
names(ras_mean_sept_1997_2012) <- c('X2012-09-16')

ras_mean_sept_1997_2012_df <- ras_mean_sept_1997_2012 %>% ras2df()

## 2) sept 2023 
names(ras_sept_2023) <- c('X2023-09-16')
ras_sept_2023_df <- ras_sept_2023 %>% ras2df()

# 3) all septs 1997-2023
names(ras_mean_sept_1997_2023) <- c('X2023-09-16')
ras_mean_sept_1997_2023_df <- ras_mean_sept_1997_2023 %>% ras2df()

# 4) sept 2005 - 2006 --- added this one in for tracks subset exploration
names(ras_mean_sept_2005_2006) <- c('X2006-09-16')
ras_mean_sept_2005_2006_df <- ras_mean_sept_2005_2006 %>% ras2df()

```

```{r 03-ras-march-means-to-df}

## 1) march 1997-2012 
names(ras_mean_march_1997_2012) <- c('X2012-03-16')

ras_mean_march_1997_2012_df <- ras_mean_march_1997_2012 %>% ras2df()

## 2) march 2023 
names(ras_march_2023) <- c('X2023-03-16')
ras_march_2023_df <- ras_march_2023 %>% ras2df()

# 3) all march 1997-2023
names(ras_mean_march_1997_2023) <- c('X2023-03-16')
ras_mean_march_1997_2023_df <- ras_mean_march_1997_2023 %>% ras2df()

# 4) march 2005 - 2006 --- added this one in for tracks subset exploration
names(ras_mean_march_2005_2006) <- c('X2006-03-16')
ras_mean_march_2005_2006_df <- ras_mean_march_2005_2006 %>% ras2df()

```


```{r 03-ras-aug-means-to-df}

## 1) aug 1997-2012 -- DKB NOTE: Changed from 2013 to 2012, since there are no aug tracks in 2013
names(ras_mean_aug_1997_2012) <- c('X2012-08-16')

ras_mean_aug_1997_2012_df <- ras_mean_aug_1997_2012 %>% ras2df()

## 2) aug 2023 
names(ras_aug_2023) <- c('X2023-08-16')
ras_aug_2023_df <- ras_aug_2023 %>% ras2df()

# 3) all augs 1997-2023
names(ras_mean_aug_1997_2023) <- c('X2023-08-16')
ras_mean_aug_1997_2023_df <- ras_mean_aug_1997_2023 %>% ras2df()

# 4) aug 2005 - 2006 --- added this one in for tracks subset exploration
names(ras_mean_aug_2006_2007) <- c('X2007-08-16')
ras_mean_aug_2006_2007_df <- ras_mean_aug_2006_2007 %>% ras2df()

```



```{r 03-plot-sept-mean-clim}

p_sept_mean <- ggplot() +
    # add sst sept all-time clim layer
    geom_raster(data = ras_mean_sept_1997_2023_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    # scale_fill_gradientn(colours = sst_cpal[6:length(sst_cpal)],
                         # breaks=seq(8,32,2),
                         # limits = c(8,32), 
    scale_fill_gradientn(colours = sst_cpal[2:length(sst_cpal)], # FYI - use this if comparing mar/septs. Otherwise uncomment lines above and use that for ms fig
                         breaks=seq(4,30,2),
                         limits = c(3,31), 

                         na.value="transparent",
                         name = "SST (°C) \n") +
    # add baselayers
    add_baseplot(bbox) 
 
# # show plot with tcms box
# p_tzcf_sept_mean_tcms <- p_tzcf_sept_mean + add_tcms_box()


```

```{r 03-plot-sept-clim-w-tzcfs}
p_sept_mean_w_tzcfs <- p_sept_mean +
    add_contour_lines(df=ras_mean_sept_1997_2012_df, contours=c(18), col='gray80') + 
    # add_contour_lines(df=ras_mean_sept_2005_2006_df, contours=c(18), col='#FFC2D4') + 
    add_contour_lines(df=ras_sept_2023_df, contours=c(18), col='snow') + 
    resize_cbar(12)

```

```{r 03-plot-sept-clim-w-tzcfs-2005-2006}
p_sept_mean_w_tzcfs_2005_2006 <- p_sept_mean +
    # add_contour_lines(df=ras_mean_sept_1997_2012_df, contours=c(18), col='gray80') + 
    add_contour_lines(df=ras_mean_sept_2005_2006_df, contours=c(18), col='#FFC2D4') + 
    add_contour_lines(df=ras_sept_2023_df, contours=c(18), col='snow') + 
    resize_cbar(12)

```

```{r 03-plot-sept-march-clim-w-tzcfs}
p_sept_march_mean_w_tzcfs <- p_sept_mean +
    # add sept isotherms
    add_contour_lines(df=ras_mean_sept_1997_2012_df, contours=c(18), col='gray80') + 
    add_contour_lines(df=ras_sept_2023_df, contours=c(18), col='snow') + 
    
    # add march isotherms
    add_contour_lines(df=ras_mean_march_1997_2012_df, contours=c(18), col='gray40') + 
    add_contour_lines(df=ras_march_2023_df, contours=c(18), col='gray20') + 

    resize_cbar(12)

```


```{r 03-plot-mar-mean-clim}

p_march_mean <- ggplot() +
    # add sst sept all-time clim layer
    geom_raster(data = ras_mean_march_1997_2023_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    # scale_fill_gradientn(colours = sst_cpal[6:length(sst_cpal)],
                         # breaks=seq(8,32,2),
                         # limits = c(8,32), 
    scale_fill_gradientn(colours = sst_cpal[2:length(sst_cpal)],
                         breaks=seq(4,30,2),
                         limits = c(3,31), 

                         na.value="transparent",
                         name = "SST (°C) \n") +
    # add baselayers
    add_baseplot(bbox) 
 
# # show plot with tcms box
# p_tzcf_sept_mean_tcms <- p_tzcf_sept_mean + add_tcms_box()


```
```{r 03-plot-march-clim-w-tzcfs}
p_march_mean_w_tzcfs <- p_march_mean +
    add_contour_lines(df=ras_mean_march_1997_2012_df, contours=c(18), col='gray80') + 
    # add_contour_lines(df=ras_mean_march_2005_2006_df, contours=c(18), col='#FFC2D4') + 
    add_contour_lines(df=ras_march_2023_df, contours=c(18), col='snow') + 
    resize_cbar(12)

```

```{r 03-plot-march-clim-w-tzcfs-2005-2006}
p_march_mean_w_tzcfs_2005_2006 <- p_march_mean +
    # add_contour_lines(df=ras_mean_march_1997_2012_df, contours=c(18), col='gray80') + 
    add_contour_lines(df=ras_mean_march_2005_2006_df, contours=c(18), col='#FFC2D4') + 
    add_contour_lines(df=ras_march_2023_df, contours=c(18), col='snow') + 
    resize_cbar(12)

```


```{r 03-plot-aug-mean-clim}

p_aug_mean <- ggplot() +
    # add sst sept all-time clim layer
    geom_raster(data = ras_mean_aug_1997_2023_df, 
                aes(x = lon, y = lat, fill = val, interpolate = TRUE)) +
    # scale_fill_gradientn(colours = sst_cpal[6:length(sst_cpal)],
                         # breaks=seq(8,32,2),
                         # limits = c(8,32), 
    scale_fill_gradientn(colours = sst_cpal[2:length(sst_cpal)], # FYI - use this if comparing mar/septs. Otherwise uncomment lines above and use that for ms fig
                         breaks=seq(4,30,2),
                         limits = c(3,31), 

                         na.value="transparent",
                         name = "SST (°C) \n") +
    # add baselayers
    add_baseplot(bbox) 
 
# # show plot with tcms box
# p_tzcf_sept_mean_tcms <- p_tzcf_sept_mean + add_tcms_box()


```

```{r 03-calc-max-lat-avgs-cc-sept}
max_lats_historic <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 9) %>%
    # group_by(id) %>%
    group_by(id, month, year) %>%
    slice(which.max(lat)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = 'historic') %>%
    as_tibble()

max_lats_cohort1 <- daily_avg_data_cohort_1 %>%
    filter(month == 9) %>%
    # group_by(id) %>%
    group_by(id, month, year) %>%
    slice(which.max(lat)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = 'cohort1') %>%
    mutate(id = as.factor(id)) %>%
    as_tibble()

max_lats_all <- rbind(max_lats_historic, max_lats_cohort1)

```

```{r 03-calc-mean-lat-avgs-cc-sept}
mean_lats_historic <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 9) %>%
    group_by(id, month, year) %>% # fyi - grouping to include month and year, for means so that it's easy to pull out 2005-2006 subset
    # slice(which.max(lat)) %>%
    summarise(lat = mean(lat, na.rm=TRUE),
              lon = mean(lon, na.rm=TRUE)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = 'historic') %>%
    as_tibble()

mean_lats_cohort1 <- daily_avg_data_cohort_1 %>%
    filter(month == 9) %>%
    group_by(id, month, year) %>% # fyi - grouping to include month and year, for means so that it's consistent for rbind'ing
    # slice(which.max(lat)) %>%
    summarise(lat = mean(lat, na.rm=TRUE),
              lon = mean(lon, na.rm=TRUE)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = 'cohort1') %>%
    mutate(id = as.factor(id)) %>%
    as_tibble()

mean_lats_all <- rbind(mean_lats_historic, mean_lats_cohort1)

```
```{r 03-calc-mean-lat-avgs-cc-mar}
mean_lats_historic_mar <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 3) %>%
    group_by(id, month, year) %>% # fyi - grouping to include month and year, for means so that it's easy to pull out 2005-2006 subset
    # slice(which.max(lat)) %>%
    summarise(lat = mean(lat, na.rm=TRUE),
              lon = mean(lon, na.rm=TRUE)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = 'historic') %>%
    as_tibble()

# mean_lats_cohort1 <- daily_avg_data_cohort_1 %>%
#     filter(month == 9) %>%
#     group_by(id, month, year) %>% # fyi - grouping to include month and year, for means so that it's consistent for rbind'ing
#     # slice(which.max(lat)) %>%
#     summarise(lat = mean(lat, na.rm=TRUE),
#               lon = mean(lon, na.rm=TRUE)) %>%
#     mutate(lon = round(lon, 3),
#            lat = round(lat, 3)) %>%
#     mutate(group = 'cohort1') %>%
#     mutate(id = as.factor(id)) %>%
#     as_tibble()
# 
# mean_lats_all <- rbind(mean_lats_historic, mean_lats_cohort1)

```


```{r 03-plot-sept-clim-w-tzcfs-max-lats-cc-pts}
p_sept_mean_w_tzcfs_cc_max_lats <- p_sept_mean_w_tzcfs + 
    add_turtle_pts(df=max_lats_historic, col="Max Sept 1997-2012 (Historic)", size=3) +
    add_turtle_pts(df=max_lats_cohort1, col="Max Sept 2023 (Cohort 1)", size=3) +
    
   scale_color_manual("Max Turtle Location in September", 
                  breaks = c("Max Sept 1997-2012 (Historic)", 
                             "Max Sept 2023 (Cohort 1)"#,
                             # "Max Sept 2005-2006 (Historic)", "Mean Sept 2005-2006 (Historic)"
                             ),
                  values = c("Max Sept 1997-2012 (Historic)"="#6D4C94",
                             "Max Sept 2023 (Cohort 1)"="#276961"#,
                             # "Max Sept 2005-2006 (Historic)"="red", "Mean Sept 2005-2006 (Historic)"="pink"
                             )) +
    theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
        # resize_cbar(24) 

p_sept_mean_w_tzcfs_cc_max_lats_w_text <- 
    p_sept_mean_w_tzcfs_cc_max_lats + 
    labs(title = "Long-term Average of SST for September (1997-2023) with Max Turtle Latitudes and TZCF Location",
    subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), their maximum lat and long locations are shown for September. \n\nThe historic tracks (1997-2012, n={nrow(mean_lats_historic)}) are shown as the purle circles. The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_historic)}), are shown as the green circles. \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    
### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-max-lats-cc-pts.png'), plot = p_sept_mean_w_tzcfs_cc_max_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```


```{r 03-plot-sept-clim-w-tzcfs-mean-lats-cc-pts}
p_sept_mean_w_tzcfs_cc_mean_lats <- p_sept_mean_w_tzcfs + 
    add_turtle_pts(df=mean_lats_historic, col="Mean Sept 1997-2012 (Historic)", size=3) +
    add_turtle_pts(df=mean_lats_cohort1, col="Mean Sept 2023 (Cohort 1)", size=3) +
        
   scale_color_manual("Mean Turtle Location in September", 
                      breaks = c("Mean Sept 1997-2012 (Historic)", 
                                 "Mean Sept 2023 (Cohort 1)"#,
                                 # "Max Sept 2005-2006 (Historic)", "Mean Sept 2005-2006 (Historic)"
                                 ),
                      values = c("Mean Sept 1997-2012 (Historic)"="mediumorchid", 
                                 "Mean Sept 2023 (Cohort 1)"="#66CCC0"
                                 )) +
        theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
    # resize_cbar(24) 

p_sept_mean_w_tzcfs_cc_mean_lats_w_text <- 
    p_sept_mean_w_tzcfs_cc_mean_lats + 
    labs(title = "Long-term Average of SST for September (1997-2023) with Monthly Average of Turtle Location and TZCF Location",
    subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), their average lat and long locations are shown for September. \n\nThe historic tracks (1997-2012, n={nrow(mean_lats_historic)}) are shown as the purle circles. The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_historic)}), are shown as the green circles. \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    

### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-mean-lats-cc-pts.png'), plot = p_sept_mean_w_tzcfs_cc_mean_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```


```{r 03-plot-sept-clim-w-tzcfs-mean-vs-max-lats-cc-pts}

p_sept_mean_w_tzcfs_cc_mean_vs_max_lats <- p_sept_mean_w_tzcfs + 
    add_turtle_pts(df=mean_lats_historic, col="Mean Sept 1997-2012 (Historic)", size=3) +
    add_turtle_pts(df=mean_lats_cohort1, col="Mean Sept 2023 (Cohort 1)", size=3) +

    add_turtle_pts(df=max_lats_historic, col="Max Sept 1997-2012 (Historic)", size=3) +
    add_turtle_pts(df=max_lats_cohort1, col="Max Sept 2023 (Cohort 1)", size=3) +

    # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2005 & year <= 2006), col="Mean Sept 2005-2006 (Historic)", size=4) +
    # add_turtle_pts(df=max_lats_historic %>% filter(year >=2005 & year <= 2006), col="Max Sept 2005-2006 (Historic)", size=4) +
    
   scale_color_manual("Max vs Mean Turtle \nLocation in September", 
                      breaks = c("Max Sept 1997-2012 (Historic)", "Mean Sept 1997-2012 (Historic)", 
                                 "Max Sept 2023 (Cohort 1)","Mean Sept 2023 (Cohort 1)"#,
                                 # "Max Sept 2005-2006 (Historic)", "Mean Sept 2005-2006 (Historic)"
                                 ),
                      values = c("Max Sept 1997-2012 (Historic)"="#6D4C94", "Mean Sept 1997-2012 (Historic)"="mediumorchid", 
                                 "Max Sept 2023 (Cohort 1)"="#276961","Mean Sept 2023 (Cohort 1)"="#66CCC0"#,
                                 # "Max Sept 2005-2006 (Historic)"="red", "Mean Sept 2005-2006 (Historic)"="pink"
                                 )) +
        theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
    # resize_cbar(24) 

p_sept_mean_w_tzcfs_cc_mean_vs_max_lats_w_text <- 
    p_sept_mean_w_tzcfs_cc_mean_vs_max_lats + 
    labs(title = "Long-term Average of SST for September (1997-2023) with Monthly Average of Turtle Location and TZCF Location",
    subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), the their maximum and average lat and long locations are shown for September. \n\nThe historic tracks (1997-2012, n={nrow(mean_lats_historic)}) are shown as the purle circles (dark purple = max, light purple = average monthly location). The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_historic)}), are shown as the green circles (dark green = max, light green = average monthly location). \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    

### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-mean-vs-max-lats-cc-pts.png'), plot = p_sept_mean_w_tzcfs_cc_mean_vs_max_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```

```{r 03-plot-sept-clim-w-tzcfs-2005-2006-mean-lats-cc-pts}

df_historic_2005_2006 <- mean_lats_historic %>% filter(year >=2005 & year <= 2006)

p_sept_mean_w_tzcfs_cc_2005_2006_lats <- p_sept_mean_w_tzcfs_2005_2006 + 
    # add_turtle_pts(df=mean_lats_historic, col="Mean Sept 1997-2012 (Historic)", size=4) +
    add_turtle_pts(df=mean_lats_cohort1, col="Mean Sept 2023 (Cohort 1)", size=3) +

    # add_turtle_pts(df=max_lats_historic, col="Max Sept 1997-2012 (Historic)", size=4) +
    # add_turtle_pts(df=max_lats_cohort1, col="Max Sept 2023 (Cohort 1)", size=3) +

    add_turtle_pts(df=mean_lats_historic %>% filter(year >=2005 & year <= 2006), col="Mean Sept 2005-2006 (Historic)", size=3) +
    # add_turtle_pts(df=max_lats_historic %>% filter(year >=2005 & year <= 2006), col="Max Sept 2005-2006 (Historic)", size=3) +
    
       scale_color_manual("Max Turtle \nLocation in September", 
                          breaks = c(#"Max Sept 1997-2012 (Historic)", "Mean Sept 1997-2012 (Historic)", 
                                     # "Max Sept 2023 (Cohort 1)",
                                     "Mean Sept 2023 (Cohort 1)",
                                     # "Max Sept 2005-2006 (Historic)"#,
                                     "Mean Sept 2005-2006 (Historic)"
                                     ),
                          values = c(#"Max Sept 1997-2012 (Historic)"="#6D4C94", "Mean Sept 1997-2012 (Historic)"="mediumorchid", 
                                     # "Max Sept 2023 (Cohort 1)"="#276961",
                                     "Mean Sept 2023 (Cohort 1)"="#66CCC0",
                                     # "Max Sept 2005-2006 (Historic)"="darkred"#,
                                     "Mean Sept 2005-2006 (Historic)"="#E05780"
                                     )) +
       theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
  # resize_cbar(24) 

# p_sept_mean_w_tzcfs_cc_2005_2006_lats_w_text <- 
#     p_sept_mean_w_tzcfs_cc_2005_2006_lats + 
#     labs(title = "Long-term Average of SST for September (1997-2023) with Monthly Average of Turtle Location and TZCF Location",
#     subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), their maximum and average lat and long locations are shown for September. \n\nThe historic tracks (2005-2006, n={nrow(df_historic_2005_2006)}) are shown as the red circles (dark red = max, light red = average monthly location). The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_cohort1)}), are shown as the green circles (dark green = max, light green = average monthly location). \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    

### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-2005-2006-mean-lats-cc-pts.png'), plot = p_sept_mean_w_tzcfs_cc_2005_2006_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```
```{r 03-plot-sept-clim-w-tzcfs-2005-2006-max-lats-cc-pts}

df_historic_2005_2006 <- max_lats_historic %>% filter(year >=2005 & year <= 2006)

p_sept_max_w_tzcfs_cc_2005_2006_lats <- p_sept_mean_w_tzcfs_2005_2006 + 
    # add_turtle_pts(df=mean_lats_historic, col="Mean Sept 1997-2012 (Historic)", size=4) +
    # add_turtle_pts(df=mean_lats_cohort1, col="Mean Sept 2023 (Cohort 1)", size=3) +

    # add_turtle_pts(df=max_lats_historic, col="Max Sept 1997-2012 (Historic)", size=4) +
    add_turtle_pts(df=max_lats_cohort1, col="Max Sept 2023 (Cohort 1)", size=3) +

    # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2005 & year <= 2006), col="Mean Sept 2005-2006 (Historic)", size=3) +
    add_turtle_pts(df=max_lats_historic %>% filter(year >=2005 & year <= 2006), col="Max Sept 2005-2006 (Historic)", size=3) +
    
       scale_color_manual("Max Turtle \nLocation in September", 
                          breaks = c(#"Max Sept 1997-2012 (Historic)", "Mean Sept 1997-2012 (Historic)", 
                                     "Max Sept 2023 (Cohort 1)",
                                     # "Mean Sept 2023 (Cohort 1)",
                                     "Max Sept 2005-2006 (Historic)"#,
                                     # "Mean Sept 2005-2006 (Historic)"
                                     ),
                          values = c(#"Max Sept 1997-2012 (Historic)"="#6D4C94", "Mean Sept 1997-2012 (Historic)"="mediumorchid", 
                                     "Max Sept 2023 (Cohort 1)"="#276961",
                                     # "Mean Sept 2023 (Cohort 1)"="#66CCC0",
                                     "Max Sept 2005-2006 (Historic)"="darkred"#,
                                     # "Mean Sept 2005-2006 (Historic)"="#E05780"
                                     )) +
       theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
  # resize_cbar(24) 

# p_sept_mean_w_tzcfs_cc_2005_2006_lats_w_text <- 
#     p_sept_mean_w_tzcfs_cc_2005_2006_lats + 
#     labs(title = "Long-term Average of SST for September (1997-2023) with Monthly Average of Turtle Location and TZCF Location",
#     subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), their maximum and average lat and long locations are shown for September. \n\nThe historic tracks (2005-2006, n={nrow(df_historic_2005_2006)}) are shown as the red circles (dark red = max, light red = average monthly location). The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_cohort1)}), are shown as the green circles (dark green = max, light green = average monthly location). \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    

### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-2005-2006-max-lats-cc-pts.png'), plot = p_sept_max_w_tzcfs_cc_2005_2006_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```


```{r 03-plot-sept-clim-w-tzcfs-2005-2006-mean-max-cc-pts}

df_historic_2005_2006 <- max_lats_historic %>% filter(year >=2005 & year <= 2006)

p_sept_mean_w_tzcfs_cc_2005_2006_lats <- p_sept_mean_w_tzcfs_2005_2006 + 
    # add_turtle_pts(df=mean_lats_historic, col="Mean Sept 1997-2012 (Historic)", size=4) +
    add_turtle_pts(df=mean_lats_cohort1, col="Mean Sept 2023 (Cohort 1)", size=3) +

    # add_turtle_pts(df=max_lats_historic, col="Max Sept 1997-2012 (Historic)", size=4) +
    add_turtle_pts(df=max_lats_cohort1, col="Max Sept 2023 (Cohort 1)", size=3) +

    add_turtle_pts(df=mean_lats_historic %>% filter(year >=2005 & year <= 2006), col="Mean Sept 2005-2006 (Historic)", size=3) +
    add_turtle_pts(df=max_lats_historic %>% filter(year >=2005 & year <= 2006), col="Max Sept 2005-2006 (Historic)", size=3) +
    
       scale_color_manual("Max vs Mean Turtle \nLocation in September", 
                          breaks = c(#"Max Sept 1997-2012 (Historic)", "Mean Sept 1997-2012 (Historic)", 
                                     "Max Sept 2023 (Cohort 1)","Mean Sept 2023 (Cohort 1)",
                                     "Max Sept 2005-2006 (Historic)", "Mean Sept 2005-2006 (Historic)"
                                     ),
                          values = c(#"Max Sept 1997-2012 (Historic)"="#6D4C94", "Mean Sept 1997-2012 (Historic)"="mediumorchid", 
                                     "Max Sept 2023 (Cohort 1)"="#276961","Mean Sept 2023 (Cohort 1)"="#66CCC0",
                                     "Max Sept 2005-2006 (Historic)"="darkred", "Mean Sept 2005-2006 (Historic)"="#E05780"
                                     )) +
       theme(legend.direction = "vertical", legend.box = "vertical") +
    resize_cbar(18) 
  # resize_cbar(24) 

p_sept_mean_w_tzcfs_cc_2005_2006_lats_w_text <- 
    p_sept_mean_w_tzcfs_cc_2005_2006_lats + 
    labs(title = "Long-term Average of SST for September (1997-2023) with Monthly Average of Turtle Location and TZCF Location",
    subtitle = glue("\n\nFor each turtle located between (165°W-140°W longitude), their maximum and average lat and long locations are shown for September. \n\nThe historic tracks (2005-2006, n={nrow(df_historic_2005_2006)}) are shown as the red circles (dark red = max, light red = average monthly location). The STRETCH Cohort 1 turtle locations (September 2023, n={nrow(mean_lats_cohort1)}), are shown as the green circles (dark green = max, light green = average monthly location). \n\nThe 18°C isotherm is a reference for the seasonal location of the TZCF. The gray line represents the average September TZCF location from 1997-2012. The white line represents the TZCF location for September 2023. \n\n"))
    

### Save plot ----
ggsave(filename = str_c('03rmd-Fig2-plot-sept-clim-w-tzcfs-2005-2006-mean-max-cc-pts.png'), plot = p_sept_mean_w_tzcfs_cc_2005_2006_lats, 
       # height=9,
       width=11, bg='white',
       path = '../figs/')
```


```{r try legends}
# horizontal_g <- p_sept_mean_w_tzcfs_cc_2005_2006_lats +  guides(color = guide_legend(order = 0),
#          color  = guide_legend(order = 1)) + theme(legend.title.align = 0.5,
#           legend.direction = "horizontal", 
#           legend.position = "bottom")
# 
# library(cowplot)
# 
# horizontal_g_legend <- horizontal_g + theme(legend.direction = "horizontal", legend.box = "horizontal") |> get_legend()
# 
# horiz_legend_only <- horizontal_g_legend %>% plot_grid(., ncol = 2, rel_widths = c(1,1))
# 
# a = horizontal_g + theme(legend.position="bottom")
# # b = horizontal_g
# 
# plot_grid(a, horiz_legend_only, ncol=2, rel_widths = c(3,0.5))

```

```{r 03-load-chlmday-nc-sept-mars}
nc_path = "/Users/briscoedk/dbriscoe@stanford.edu - Google Drive/My Drive/ncdf/npac"

chl_mday_ncs_sept <- list.files(nc_path, pattern = str_c("chla_erdMH1chlamday_Lon0360_"), full.names=T) %>%
    .[grep("-09-", .)]

chl_mday_ncs_mar <- list.files(nc_path, pattern = str_c("chla_erdMH1chlamday_Lon0360_"), full.names=T) %>%
    .[grep("-03-", .)]


sept_fdates <- chl_mday_ncs_sept %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 4)) %>%
    substr(., start=1, stop=10)

mar_fdates <- chl_mday_ncs_mar %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 4)) %>%
    substr(., start=1, stop=10)


chl_mday_ras_stack_sept <- chl_mday_ncs_sept %>% #ncs[fdates >= params$min_dt & fdates <= params$max_dt] %>%
    raster::stack(.)

chl_mday_ras_stack_mar <- chl_mday_ncs_mar %>% #ncs[fdates >= params$min_dt & fdates <= params$max_dt] %>%
    raster::stack(.)

```

```{r 03-calc-ras-means-chlamday-sept-mars-2003-2011}

# set Z - Sept -------
chl_mday_ras_stack_sept <- setZ(chl_mday_ras_stack_sept, sept_fdates)
names(chl_mday_ras_stack_sept) <- as.character(sept_fdates)  # mon-yr abbreb

# 3006-2007 Sept subset
chl_mday_ras_stack_sept_2006_2007 <- subset(chl_mday_ras_stack_sept, c(4:5))
names(chl_mday_ras_stack_sept_2006_2007) <- as.character(sept_fdates)[4:5]  # mon-yr abbreb

# set Z - Mar
chl_mday_ras_stack_mar <- setZ(chl_mday_ras_stack_mar, mar_fdates)
names(chl_mday_ras_stack_mar) <- as.character(mar_fdates)  # mon-yr abbreb


## CALC RAS MEAN -------
# Septs
chla_ras_mean_2003_2011_sept <- chl_mday_ras_stack_sept %>%
     calc(., mean, na.rm=TRUE)

# Septs 2006-2007
chl_mday_ras_mean_sept_2006_2007 <- chl_mday_ras_stack_sept_2006_2007 %>%
     calc(., mean, na.rm=TRUE)

# Mars
chla_ras_mean_2003_2011_mar <- chl_mday_ras_stack_mar %>%
     calc(., mean, na.rm=TRUE)


# RESET MEAN RAS NAMES ----
# Septs
names(chla_ras_mean_2003_2011_sept) = as.character(getZ(chl_mday_ras_stack_sept) |> (\(x) { tail(x, n=1)})() )# assign dummy date for mean ras. use last 

names(chl_mday_ras_mean_sept_2006_2007) = as.character(getZ(chl_mday_ras_stack_sept_2006_2007) |> (\(x) { tail(x, n=1)})() )

# Mars
names(chla_ras_mean_2003_2011_mar) = as.character(getZ(chl_mday_ras_stack_mar) |> (\(x) { tail(x, n=1)})() )# assign dummy date for mean ras. use last date


# CONVERT TO DF -------
# Septs
chla_ras_mean_2003_2011_sept_df <- chla_ras_mean_2003_2011_sept  %>% ras2df() %>%
    mutate(lon= make180(lon))

chl_mday_ras_mean_sept_2006_2007_df <- chl_mday_ras_mean_sept_2006_2007  %>% ras2df() %>%
    mutate(lon= make180(lon))

# Mars
chla_ras_mean_2003_2011_mar_df <- chla_ras_mean_2003_2011_mar  %>% ras2df() %>%
    mutate(lon= make180(lon))


# chl_mday_ras_stack_sept_df <- chl_mday_ras_stack_sept 
```

```{r 03-set-chla-cpal-blue-green-zCuts}
# CBAR
myPallette <-
  c(rev(brewer.pal(8, "Greens")[2:8])
    # , "white"
    , brewer.pal(3, "Blues")[2:2])

zCuts <- c(0,0.2, 0.3, 0.5,1.0,2,5,25)


```


```{r 03-gg-sept-chl-with-cc-sst-isotherm-w-uturns}

## PLOT (with u-turn lower lat tags)

p_sept_mean_chla_2003_2011_uturns <- ggplot(chla_ras_mean_2003_2011_sept_df %>% drop_na(), aes(lon, lat)) +
  # geom_raster(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     geom_tile(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
  scale_fill_manual("Mean Chla", na.translate = F,
                    values = rev(myPallette), na.value="transparent",
                      labels = c("0","0.2", "0.3", "0.5","1.0","2",">5",">25")
                      ) + 
    add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(18), col='white') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(17), col='azure4') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(19.5), col='black') +
    add_baseplot(bbox) +
    guides(fill = guide_legend(title = 'Mean Sept\nChla (mg/m^3)')) + 
    # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2006 & year <= 2007), col="2006-2007", size=3) +
     add_turtle_pts(df=mean_lats_historic %>% filter(year >=2003 & year <= 2011), col="2003-2011", size=3) +
     # add_turtle_pts(df=subset_mn_lats %>% filter(year >= 2002 & year <= 2011), col="2002-2011", size=3) +
    
       scale_color_manual("Mean Sept\nTurtle Location", 
                          breaks = c(
                              "2003-2011"#,
                                     # "2006-2007"
                                     ),
                          values = c(
                              "2003-2011"="lightslateblue"#,
                                     # "2006-2007"="green"
                                     )) +
        labs(subtitle = "Mean Sept Chl-a (2003-2011) and Turtle Location (2003-2011)",
         caption = str_c("Isotherms: 17°C (gray line), 18°C (white line), 19.5°C (black line)")) + 
         xlab("\nLongitude") +
         ylab("Latitude\n")

    
    ### Save plot ----
ggsave(filename = str_c('plot_sept_mean_chla_turtles_w_sst_isotherms_2003_2011_w_uturns.png'), plot = p_sept_mean_chla_2003_2011_uturns, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

```

```{r 03-gg-sept-chl-with-cc-sst-isotherm-rm-uturns}

## MN FILT (remove u-turns!)
subset_mn_lats <- mean_lats_historic %>%
    filter(!(id == "50136_05" & year == "2007")) %>%
    filter(!(id == "57142_05" & year == "2006")) %>%
     filter(!(id == "23177_05" & year == "2007")) %>%
     filter(!(id == "23045_03" & year == "2006")) 

## PLOT (w/o u-turn tags)
p_sept_mean_chla_2003_2011_rm_uturns <- ggplot(chla_ras_mean_2003_2011_sept_df %>% drop_na(), aes(lon, lat)) +
  # geom_raster(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     geom_tile(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
  scale_fill_manual("Mean Chla", na.translate = F,
                    values = rev(myPallette), na.value="transparent",
                      labels = c("0","0.2", "0.3", "0.5","1.0","2",">5", ">25")
                      ) + 
    add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(18), col='white') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(17), col='azure4') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(19.5), col='black') +
    add_contour_lines(df=ras_sept_2023_df, contours=c(18), col='pink') +
    
    add_baseplot(bbox) +
    guides(fill = guide_legend(title = 'Mean Sept\nChla (mg/m^3)')) + 

     # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2002 & year <= 2011), col="2002-2011", size=3) +
     add_turtle_pts(df=subset_mn_lats %>% filter(year >= 2003 & year <= 2011), col="2003-2011", size=3) +
     add_turtle_pts(df=mean_lats_cohort1, col="2023", size=3) +
    
       scale_color_manual("Mean Sept\nTurtle Location", 
                          breaks = c(
                              "2003-2011",
                                     "2023"
                                     ),
                          values = c(
                              "2003-2011"="mediumslateblue",
                                     "2023"="#E05087"
                                     )) +
        labs(subtitle = "Mean Sept Chl-a (2003-2011) and Turtle Location (2003-2011)",
         caption = str_c("Isotherms: 17°C (gray line); 18°C (2003-2011, white line), 18°C (2023, pink line); 19.5°C (black line)")) + 
         xlab("\nLongitude") +
         ylab("Latitude\n")


## Save plots
ggsave(filename = str_c('plot_sept_mean_chla_turtles_w_sst_isotherms_2003_2011_rm_uturns_w_2023.png'), plot = p_sept_mean_chla_2003_2011_rm_uturns, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )
    
```


```{r 03-gg-mar-chl-with-cc-sst-isotherm-w-uturns}

## PLOT (with u-turn lower lat tags)

p_mar_mean_chla_2003_2011_uturns <- ggplot(chla_ras_mean_2003_2011_mar_df %>% drop_na(), aes(lon, lat)) +
  # geom_raster(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     geom_tile(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     scale_fill_manual("Mean Chla", na.translate = F,
                      values = rev(myPallette), na.value="transparent",
                      labels = c("0","0.2", "0.3", "0.5","1.0","2",">5",">25")
                      ) + 
    add_contour_lines(df=ras_mean_march_1997_2012_df %>% filter(date >= "2003-03-01" | date <= "2011-03-31"), contours=c(18), col='white') +
        add_contour_lines(df=ras_mean_march_1997_2012_df %>% filter(date >="2003-03-01" | date <= "2011-03-31"), contours=c(17), col='azure4') +
        add_contour_lines(df=ras_mean_march_1997_2012_df %>% filter(date >= "2003-03-01" | date <= "2011-03-31"), contours=c(19.5), col='black') +
     add_contour_lines(df=ras_mean_march_1997_2012_df %>% filter(date >= "2003-03-01" | date <= "2011-03-31"), contours=c(16), col='pink') +
    add_baseplot(bbox) +
    guides(fill = guide_legend(title = 'Mean March\nChla (mg/m^3)')) + 
    # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2006 & year <= 2007), col="2006-2007", size=3) +
     add_turtle_pts(df=mean_lats_historic_mar %>% filter(year >=2003 & year <= 2011), col="2003-2011", size=3) +
     # add_turtle_pts(df=subset_mn_lats %>% filter(year >= 2002 & year <= 2011), col="2002-2011", size=3) +
    
       scale_color_manual("Mean March\nTurtle Location", 
                          breaks = c(
                              "2003-2011"#,
                                     # "2006-2007"
                                     ),
                          values = c(
                              "2003-2011"="lightslateblue"#,
                                     # "2006-2007"="green"
                                     )) +
        labs(subtitle = "Mean March Chl-a (2003-2011) and Turtle Location (2003-2011)",
         caption = str_c("Isotherms: 16°C (pink line), 17°C (gray line), 18°C (white line), 19.5°C (black line)")) + 
         xlab("\nLongitude") +
         ylab("Latitude\n")

    
    ### Save plot ----
ggsave(filename = str_c('plot_mar_mean_chla_turtles_w_sst_isotherms_2003_2011_w_uturns.png'), plot = p_mar_mean_chla_2003_2011_uturns, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

```

```{r 03-gg-sept-chl-with-cc-aug-sst-isotherm-rm-uturns}
# PLOT OF SEPT CHLA & CC TAGS with AUG SST ISOTHERMS -----------

## MN FILT (remove u-turns!)
subset_mn_lats <- mean_lats_historic %>%
    filter(!(id == "50136_05" & year == "2007")) %>%
    filter(!(id == "57142_05" & year == "2006")) %>%
     filter(!(id == "23177_05" & year == "2007")) %>%
     filter(!(id == "23045_03" & year == "2006")) 

## PLOT (w/o u-turn tags)
p_sept_mean_chla_aug_sst_2003_2011_2023_rm_uturns <- ggplot(chla_ras_mean_2003_2011_sept_df %>% drop_na(), aes(lon, lat)) +
  # geom_raster(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     geom_tile(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
  scale_fill_manual("Mean Chla", na.translate = F,
                    values = rev(myPallette), na.value="transparent",
                      labels = c("0","0.2", "0.3", "0.5","1.0","2",">5", ">25")
                      ) + 
     
 add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(18), col='azure4') +
        # add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(17), col='azure4') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(19.5), col='black') +
 #    
    add_contour_lines(df=ras_aug_2023_df, contours=c(18), col='pink') +
    
     # add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(18), col='azure4') +
     #    # add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(17), col='azure4') +
     #    add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(19.5), col='black') +
  
    
              geom_contour(data=ras_mean_aug_1997_2012_df %>% filter(date >= "2003-08-01" | date <= "2011-08-31"), aes(x=lon, y=lat, z = val), linetype = "dotted",
              colour = 'azure4', linewidth =1, breaks = c(18), show.legend = F) +
    
          geom_contour(data=ras_mean_aug_1997_2012_df %>% filter(date >= "2003-08-01" | date <= "2011-08-31"), aes(x=lon, y=lat, z = val), linetype = "dotted",
              colour = 'gray10', linewidth = 0.65, breaks = c(19.5), show.legend = F) +
    
    add_baseplot(bbox) +
    guides(fill = guide_legend(title = 'Mean Sept\nChla (mg/m^3)')) + 

     # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2002 & year <= 2011), col="2002-2011", size=3) +
     add_turtle_pts(df=subset_mn_lats %>% filter(year >= 2003 & year <= 2011), col="2003-2011", size=3) +
     add_turtle_pts(df=mean_lats_cohort1, col="2023", size=3) +
    
       scale_color_manual("Mean Sept\nTurtle Location", 
                          breaks = c(
                              "2003-2011",
                                     "2023"
                                     ),
                          values = c(
                              "2003-2011"="mediumslateblue",
                                     "2023"="#E05087"
                                     )) +
        labs(subtitle = "Mean Sept Chl-a (2003-2011) and Turtle Location (2003-2011)",
         caption = str_c("Isotherms: Sept 18°C (2003-2011, gray line), Sept 19.5°C (2003-2011, black line)\n 
         August 18°C (dotted gray line), August 19.5°C (dotted black line)\n
                         Sept 18°C (2023, pink line)")) + 
         xlab("\nLongitude") +
         ylab("Latitude\n")


## Save plots
ggsave(filename = str_c('plot_sept_mean_chla_turtles_w_aug_sst_isotherms_2003_2011_and_2023.png'), plot = p_sept_mean_chla_aug_sst_2003_2011_2023_rm_uturns, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )
    
```

```{r 03-gg-sept-chl-with-cc-2006-2007-2023-aug-sst-isotherm-rm-uturns}
# PLOT OF SEPT CHLA & CC TAGS with AUG SST ISOTHERMS -----------

## MN FILT (remove u-turns!)
subset_mn_lats <- mean_lats_historic %>%
    filter(!(id == "50136_05" & year == "2007")) %>%
    filter(!(id == "57142_05" & year == "2006")) %>%
     filter(!(id == "23177_05" & year == "2007")) %>%
     filter(!(id == "23045_03" & year == "2006")) 

## PLOT (w/o u-turn tags)
p_sept_mean_chla_aug_sst_2006_2007_2023_rm_uturns <- ggplot(chl_mday_ras_mean_sept_2006_2007_df %>% drop_na(), aes(lon, lat)) +
  # geom_raster(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
     geom_tile(aes(x = lon, y = lat, fill = cut(val, zCuts), interpolate = TRUE, na.rm = TRUE)) +
  scale_fill_manual("Mean Chla", na.translate = F,
                    values = rev(myPallette), na.value="transparent",
                      labels = c("0","0.2", "0.3", "0.5","1.0","2",">5", ">25")
                      ) + 
     
 add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2006-09-01" | date <= "2007-09-30"), contours=c(18), col='azure4') +
        # add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2003-09-01" | date <= "2011-09-30"), contours=c(17), col='azure4') +
        add_contour_lines(df=ras_mean_sept_1997_2012_df %>% filter(date >= "2006-09-01" | date <= "2007-09-30"), contours=c(19.5), col='black') +
 #    
    add_contour_lines(df=ras_aug_2023_df, contours=c(18), col='pink') +
    
     # add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(18), col='azure4') +
     #    # add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(17), col='azure4') +
     #    add_contour_lines_dashed(df=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), contours=c(19.5), col='black') +
  
    
              geom_contour(data=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), aes(x=lon, y=lat, z = val), linetype = "dotted",
              colour = 'azure4', linewidth =1, breaks = c(18), show.legend = F) +
    
          geom_contour(data=ras_mean_aug_1997_2012_df %>% filter(date >= "2006-08-01" | date <= "2007-08-31"), aes(x=lon, y=lat, z = val), linetype = "dotted",
              colour = 'gray10', linewidth = 0.65, breaks = c(19.5), show.legend = F) +
    
    add_baseplot(bbox) +
    guides(fill = guide_legend(title = 'Mean Sept\nChla (mg/m^3)')) + 

     # add_turtle_pts(df=mean_lats_historic %>% filter(year >=2002 & year <= 2011), col="2002-2011", size=3) +
     add_turtle_pts(df=subset_mn_lats %>% filter(year >= 2006 & year <= 2007), col="2006-2007", size=3) +
     add_turtle_pts(df=mean_lats_cohort1, col="2023", size=3) +
    
       scale_color_manual("Mean Sept\nTurtle Location", 
                          breaks = c(
                              "2006-2007",
                                     "2023"
                                     ),
                          values = c(
                              "2006-2007"="mediumslateblue",
                                     "2023"="#E05087"
                                     )) +
        labs(subtitle = "Monthly Mean Sept Chl-a (2006-2007) and Turtle Location (2006-2007, 2023)",
         caption = str_c("Isotherms: Sept 18°C (2006-2007, gray line), Sept 19.5°C (2006-2007, black line)\n 
         August 18°C (2006-2007, dotted gray line), August 19.5°C (2006-2007, dotted black line)\n
                         Sept 18°C (2023, pink line)")) + 
         xlab("\nLongitude") +
         ylab("Latitude\n")


## Save plots
ggsave(filename = str_c('plot_sept_mean_chla_turtles_w_aug_sst_isotherms_2006_2007_and_2023.png'), plot = p_sept_mean_chla_aug_sst_2006_2007_2023_rm_uturns, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )
    
```




```{r 03-interim-rdata-save}

## INTERIM SAVE 
f <- "../data/interim/03_northern_lats_EDA_interim_data.RData"
if (file.exists(f)){
    invisible()
} else{
    save.image(file = "../data/interim/03_northern_lats_EDA_interim_data.RData")
}

```


<!-- ```{r} -->

<!-- p_sept_mean_w_tzcfs_cc_max_lats + -->
<!--     geom_point(data = max_lats_historic, aes(x = lon, y = lat, color = "#6D4C94"), size = 3, shape = 16, stroke = 1) + -->

<!--   # Add points from the second dataset -->
<!--   geom_point(data = max_lats_cohort1, aes(x = lon, y = lat), color = "#66CCC0", size = 3, shape = 16, stroke = 1) + -->

<!--   # Customize the legend -->
<!--   scale_color_identity(guide = "legend", labels = c("Dataset 1", "Dataset 2")) -->


<!-- geom_point(data = max_lats_historic, aes(x = lon, y = lat), size = 3, shape = 16, stroke = 1, color = "black") + -->
<!-- geom_point(data = max_lats_cohort1, aes(x = lon, y = lat), size = 3, shape = 16, stroke = 1, color = "black") -->



<!-- p_sept_mean_w_tzcfs +  -->
<!--     add_turtle_pts(df=max_lats_historic, col="Sept 1997-2013") + -->
<!--     add_turtle_pts(df=max_lats_cohort1, col="Sept 2023") + -->
<!--       # geom_point(data = g1, aes(colour = "Point 15", shape = "Point 15"), size = 4) + -->
<!--      scale_shape_manual(values = c(16, 21) ) + -->
<!--     # scale_fill_manual(values = c("pink", "turquoise") ) + -->
<!--      scale_color_manual(values = c("white", "white") ) + -->
<!--      guides(color = guide_legend( override.aes = list(size = c(3,3) ) ) ) -->



<!--     # ggplot() + -->
<!--     geom_point(data=max_lats_all, aes(x = lon, y = lat), fill="green",  -->
<!--                color = "white", -->
<!--                 stroke = 1, -->
<!--                 shape = 21, -->
<!--                 size=3) + -->

<!--      geom_point(data = grid_data, aes(x = grid_data$long, y = grid_data$lat), size = 3,  -->
<!--         shape = 23, fill = "orange") -->

<!--     scale_fill_manual( -->
<!--         name = "Max Turtle Latitudes", -->
<!--         values = c("#66CCC0", "#6D4C94") -->
<!--     ) + -->

<!--     resize_cbar(16)  -->

<!-- p_sept_mean_w_tzcfs +  -->
<!--  geom_point(color='black', shape=21, size=4, aes(fill=factor(group))) +  -->
<!--   scale_fill_manual(values=c('pink', 'lightgreen')) -->


<!-- scale_color_manual( -->
<!--                      values=c("black", "black")) + -->
<!--   scale_fill_manual(values=c("#6D4C94","#66CCC0")) -->

<!-- ``` -->
