==> rmd/04_northern_lats_EDA_tzcf_figs.Rmd <==
---
title: "4 | TZCF Figs"
subtitle: "Project STRETCH: Northern Latitudes - Exploratory Analyses"
author: "Dana K Briscoe"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: none
    df_print: paged
    highlight: tango
    number_sections: no
    theme: flatly
    toc: no
    toc_float: true
    toc_collapsed: true
  pdf_document:
    toc: yes
    toc_depth: '2'
---

```{=html}
<style>
.main-container {
    max-width: 1180px;
    margin-left: 10;
    margin-right: 10;
}
</style>
```

```{r 04-setup, include=FALSE}
knitr::opts_chunk$set(
    echo = FALSE,
    eval = TRUE,
    message = FALSE,
    warning = FALSE,
    dpi = 300,
    fig.align = "center"
    )

# See Options: knitr::opts_chunk$get()
```

```{r 04-load-libraries, message=FALSE}
# 01
library(tidyverse)
library(data.table)
library(ggplot2)
library(plotly)
library(leaflet)
library(mapview)
library(leafsync)
library(leafgl)
library(sf)
library(kableExtra)
library(knitr)
# 02
library(lubridate)
library(scales)
# 03 
library(RColorBrewer)
library(fields)
library(colorRamps)
require(latticeExtra)
require(zoo)
library(raster)
#04
library(plotly)
```

```{r 04-source-helper-funcs}
source('../code/00_northern_lats_helper_functions.R')
```

```{r 04-load-interim-data}
f <- "../data/interim/04_northern_lats_EDA_interim_data.RData"
if (file.exists(f)){
    load(f)
}

```


```{r 04-load-cohort-1}
source('../code/01_prep_turtle_data.R')

```

```{r 04-rename-dfs-for-consistency}
raw_df <- raw_data_cohort_1_w_names
daily_df <- daily_avg_data_cohort_1

# make historic df consistent with cohort df(s)
hist_df <-  historic_tags %>% wrangleHistDF()

```

```{r 04-set-spatial-extent}
bbox <- c(make360(-170), make360(-110), 20, 50)
e <- extent(bbox[1], bbox[2], bbox[3], bbox[4])

tzcf_bbox <- c(make360(-160), make360(-130), 25, 50)
e_tzcf <- extent(tzcf_bbox[1], tzcf_bbox[2], tzcf_bbox[3], tzcf_bbox[4])

```

```{r 04-set-ras-params}
params <- tibble(
    eov = "sst",
    timestep = "month",
    # min_dt = '1997-07-16',
    min_dt = '1985-01-16',
    max_dt = '2023-10-16')

```


```{r 04-load sst-monthly-ncdf}

nc_path = "/Users/briscoedk/dbriscoe@stanford.edu - Google Drive/My Drive/ncdf/npac"

ncs <- list.files(nc_path, pattern = str_c(params$eov, "_"), full.names=T) %>%
    .[grep(params$timestep, .)]


fdates <- ncs %>% 
    str_split(., "_") %>%
    purrr::map_chr(~ pluck(., 6)) %>%
    substr(., start=1, stop=10)
       

ras_stack <- #ncs[fdates >= params$min_dt & fdates <= params$max_dt] %>%
    ncs %>%
    raster::stack(.) %>%
    # crop(., e)
    crop(., e_tzcf)

```


```{r 04-set-ras-info}

dt_idx <- getDateRange(params$min_dt, params$max_dt, unit = params$timestep)
ras_names_yrmon <- as.yearmon(dt_idx)

ras_stack <- setZ(ras_stack, ras_names_yrmon)
names(ras_stack) <- as.character(ras_names_yrmon)  # mon-yr abbreb
# names(ras_stack) <- as.character(dt_idx) # full Xyr-mon-d 

```

```{r 04-set-turtle-cpal}
turtle_cpal <- c("#66CCC0", "#6D4C94") # cohort1, historic
```


```{r}
# ras_avg_1997_2022 <- subset(ras_stack, which(getZ(ras_stack) <= "2022-12-16"))
# 
# ras_2023 <- subset(ras_stack, which(getZ(ras_stack) >= "2023-01-16"))
# 
# # #get the date from the names of the layers and extract the month
# # indices <- format(as.Date(names(tmp_ras), format = "X%Y.%m.%d"), format = "%m")
# # indices <- as.numeric(indices)
# 
# # dirLayer <- tmp_ras %>% 
# #     init(., v='y')            # spatial avg across lon (in this case, 20 deg chunks)
# # 
# #       z <- zonal(tmp_ras, dirLayer, FUN='mean', digits=2)
# #       # # idx<-all.mons[1:length(rbextract.ECNP@layers)]
# #        idx <- seq(as.Date('2023-01-16'), as.Date('2023-08-16'), by='month')
# # 
# #       # idx <- seq(as.Date('1997-01-16'), as.Date('2023-08-16'), by='month')
# #       # idx <- as.Date(fdates)   
# #       
# #       dat <- expand.grid(y=z[,1], x=idx)
# #       dat$z <- as.vector(z[,-1], mode='numeric')
# #       
# # dat$x <- as.Date(dat$x)
# # # head(dat)
# 
# idx_2023 = seq(as.Date('2023-01-16'), as.Date('2023-08-16'), by='month')
# dat_2023 <- getHovLon(ras_2023, idx_2023)

```

```{r 04-subset-ras-stack-by-years}
# get unique years from time attribute
years <- getZ(ras_stack) |> format("%Y") |> unique()

 
# names(tmp_stack) <- names(tmp_stack) |> stringr::str_sub(start = 2, end = 15) |> strptime(format = "%Y.%m.%d", tz = "UTC") |> format("%Y")

# function
subset_by_year <- function(year, stack) {
  subset(stack, grep(year, names(stack)))
}

# subset stack into a list by year
ras_stack_by_year_list <- lapply(years, subset_by_year, stack = ras_stack)

```


```{r 04-get-Hov-Lon-MonYr}
raster_stack <- ras_stack

# DO THIS!
# Define the start and end year and month
# start_date <- as.Date("1997-07-01")
# end_date <- as.Date("2023-09-01")
start_date <- as.Date("1985-01-16")
end_date <- as.Date("2023-10-16")
date_seq <- seq(start_date, end_date, by = "1 month")

# Create a data frame of year-month combinations for every month between the specified period
year_month_df <- data.frame(
  year = year(date_seq),
  month = month(date_seq)
)

# reset names for raser stack to run getHovLon
names(raster_stack) <- apply(year_month_df, 1, function(row) {
  sprintf("layer_%04d_%02d", row["year"], row["month"])
})

# Extract year and month from layer names
layer_names <- names(raster_stack)
year_month_info <- strsplit(layer_names, "_")
years <- as.integer(sapply(year_month_info, function(x) as.integer(x[2])))
months <- as.integer(sapply(year_month_info, function(x) as.integer(x[3])))


# Use purrr's map2 function to apply getHovLon to each month and year combination
result_list <- map2(years, months, ~ {
  year <- .x
  month <- .y
  subset_layers <- grep(paste0("^layer_", year, "_", sprintf("%02d", month)), layer_names, value = TRUE)
  subset_raster <- raster_stack[[subset_layers]]
  getHovLon(subset_raster, year, month)
})

```

```{r 04-convert-hov-mon-yr-to-df}
# Combine the list of data frames into one large data frame
combined_df <- do.call(rbind, result_list)

# Group the combined data frame by year and month
grouped_by_year_month <- combined_df %>%
    mutate(year = lubridate::year(x),
           month = lubridate::month(x),
           day = lubridate::day(x)) %>%
    mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) %>%
  group_by(year, month)

```

```{r 04-interim-save-grouped-by-year-month-tzcf-hov-df}
saveRDS(grouped_by_year_month, "../data/interim/04_grouped_by_year_month_Jan1981_Oct2023.rds")


```



```{r 04-calc-avg-tzcf-1997-2012}
grouped_by_year_tzcf_1997_2012 <- grouped_by_year_month %>%
    filter(x >= '1997-01-01' & x <= '2012-12-31') %>%
    group_by(month, y) %>%
     summarise(
               z = mean(z)) %>%
    mutate(x = as.Date(paste0("2012-", month, "-01"))) %>%
    mutate(year = lubridate::year(x),
           month = lubridate::month(x),
           day = lubridate::day(x)) %>%
    mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) 
```

```{r 04-calc-avg-tzcf-2005-2006}
grouped_by_year_tzcf_2005_2006 <- grouped_by_year_month %>%
    filter(x >= '2005-01-01' & x <= '2006-12-31') %>%
    group_by(month, y) %>%
     summarise(
               z = mean(z)) %>%
    mutate(x = as.Date(paste0("2006-", month, "-01"))) %>%
    mutate(year = lubridate::year(x),
           month = lubridate::month(x),
           day = lubridate::day(x)) %>%
    mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) 
```

```{r 04-calc-avg-tzcf-2006-2007}
grouped_by_year_tzcf_2006_2007 <- grouped_by_year_month %>%
    filter(x >= '2006-01-01' & x <= '2007-12-31') %>%
    group_by(month, y) %>%
     summarise(
               z = mean(z)) %>%
    mutate(x = as.Date(paste0("2007-", month, "-01"))) %>%
    mutate(year = lubridate::year(x),
           month = lubridate::month(x),
           day = lubridate::day(x)) %>%
    mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) 
```

```{r 04-set-enso-conds}

enso_yr <- data.frame(year = seq(1985,2023,1)) %>%
    mutate(enso = case_when((year == 1986 | year == 1987 | year == 1991 | year == 1994 | year == 1997 | 
                                 year == 2002 | year == 2004 | year == 2006 | year == 2009 | year == 2015 | year == 2018 | year == 2023) ~ "el_nino",
                            
                            (year == 1988 | year == 1995 | year == 1998 | year == 1999 | 
                                 year == 2000 | 
                                 # year == 2005 | 
                                 year == 2007 | 
                                 # year == 2008 | 
                                 year == 2010 | year == 2011 | year == 2016 | year == 2017 | 
                                 year == 2020 | year == 2021 | year == 2022) ~ "la_nina",
                            
                            (year == 2014 | year == 2015 | year == 2019 | year == 2020) ~ "mhw",    
                            TRUE ~ "neutral"))



sept_dte <- getZ(sept_sst_ras_stack)
sept_y <- format(sept_dte, "%Y")

```



```{r 04-calc-avg-tzcf-by-enso}

grouped_by_year_tzcf_enso_cond <- merge(grouped_by_year_month, enso_yr, by = "year")


grouped_by_year_tzcf_enso_all <- grouped_by_year_tzcf_enso_cond %>%
    # filter(enso == 'el_nino') %>%
    group_by(enso, month, y) %>%
     summarise(
               z = mean(z),
               z_min = min(z),
               z_sd_lwr = mean(z)-sd(z),
               z_sd_upr = mean(z)+sd(z),
               z_max = max(z)) %>%
    mutate(x = as.Date(paste0("2023-", month, "-01"))) %>%
    mutate(year = lubridate::year(x),
           month = lubridate::month(x),
           day = lubridate::day(x)) %>%
    mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) 



# # el
# grouped_by_year_tzcf_el <- grouped_by_year_tzcf_enso_all %>%
#     filter(enso == 'el_nino') %>%
#     group_by(month, y) %>%
#      # summarise(
#      #           z = mean(z)) %>%
#     mutate(x = as.Date(paste0("1997-", month, "-01"))) %>%
#     mutate(year = lubridate::year(x),
#            month = lubridate::month(x),
#            day = lubridate::day(x)) %>%
#     mutate(month_abb = month.abb[as.numeric(format(x, "%m"))]) %>%
#     mutate(enso = 'el_nino')

```




```{r 04-set-ggplot-layers}

add_baseplot_theme <- function(){
    list(
        labs(x = "Month", y = "Latitude (°N)"),
        scale_x_continuous(
        breaks = 1:12,  # Specify the breaks for months (1 to 12)
        labels = month.abb  # Use the built-in month abbreviation vector
        ),
        theme_minimal()
    )
}

add_tzcf_contour <- function(df, col='gray70'){
    list(
        geom_contour(data=df,
        aes(x=month, y=y, z = z, text = paste("Avg location "), colour = col),
        linewidth = 1.25, breaks = c(18))
    )
}

add_tzcf_indiv_contours <- function(df, col='gray70'){
    list(
        geom_contour(data=df,
        aes(x=month, y=y, z = z, text = paste("Year: ", year), colour = col),
        linewidth = 1.25, breaks = c(18))
    )
}

add_tzcf_contour_2023 <- function(col='dodgerblue'){
    list(
        geom_contour(data=grouped_by_year_month %>% filter(year == 2023),
        aes(x=month, y=y, z = z, group = as.factor(year), text = paste("Year: ", year), colour = col),
        linewidth = 1.25, breaks = c(18))
    )
}


add_historic_mlats <- function(col="#6D4C94"){
    list(
        # geom_point(data=mlats_historic, aes(x = month, y = avgs, text = str_c("Avg Lat: ", round(avgs, 2)), color=col), size = 3),
        geom_line(data=mlats_historic, aes(x = month, y = avgs, colour = col), linewidth = 1.25)
    )
}

add_cohort1_mlats <- function(col="#66CCC0"){
    list(
        # geom_point(data=mlats_cohort1, aes(x = month, y = avgs, text = str_c("Avg Lat: ", round(avgs, 2)), color=col), size = 3),
        geom_line(data=mlats_cohort1, aes(x = month, y = avgs, colour = col), linewidth = 1.25)
    )
}

add_historic_2005_2006_mlats <- function(col="#E05780"){
    list(
        # geom_point(data=mlats_cohort1, aes(x = month, y = avgs, text = str_c("Avg Lat: ", round(avgs, 2)), color=col), size = 3),
        geom_line(data=mlats_historic_2005_2006, aes(x = month, y = avgs, colour = col), linewidth = 1.25)
    )
}

add_historic_2006_2007_mlats <- function(col="#E05780"){
    list(
        # geom_point(data=mlats_cohort1, aes(x = month, y = avgs, text = str_c("Avg Lat: ", round(avgs, 2)), color=col), size = 3),
        geom_line(data=mlats_historic_2006_2007, aes(x = month, y = avgs, colour = col), linewidth = 1.25)
    )
}

```

```{r 04-plot-tzcf-all-yrs}
p_tzcf_all_yrs <-  ggplot() +
    # Create additional contour layers (replace with your data and aesthetics)
       add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year != 2023), col="Individual Years (1997-2022)") +
    
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2005), col="2005") +
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2015), col="2015") +
         add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2019), col="2019") +
         # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2020), col="2020") +
    
        add_tzcf_contour_2023(col="2023") + 
    
        add_baseplot_theme() +
    
    scale_y_continuous(breaks=seq(30, 45, 2)) +

        scale_color_manual(
        name = "18°C Isotherm (165°W-140°W)\n ",
        values = c("#E4938B", "#EFC1BD","#F8C08C", "#952C23", "gray80")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("Individual Years (1997-2022)", "2015","2023")  # Specify labels
        )


gg_tzcf_all_yrs <- ggplotly(p_tzcf_all_yrs, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))

```


```{r 04-plot-tzcf-all-yrs-2006-2007-2023-yrs}
p_tzcf_all_yrs_2006_2007_2023 <-  ggplot() +
    # Create additional contour layers (replace with your data and aesthetics)
       add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year != 2023), col="Individual Years\n(1981-2022)") +
    
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2006), col="2006") +
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2007), col="2007") +
         # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2019), col="2019") +
         # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2020), col="2020") +
    
        add_tzcf_contour_2023(col="2023") + 
    
        add_baseplot_theme() +
    
    scale_y_continuous(breaks=seq(30, 45, 2)) +

        scale_color_manual(
        name = "18°C Isotherm\n(165°W-140°W)\n ",
        values = c("#E4938B", "#F8C08C", "#952C23", "gray80")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("Individual Years (1997-2022)", "2015","2023")  # Specify labels
        )

### Save plot ----
ggsave(filename = 'plot_tzcf_all_yrs_2006_2007_2023.png', plot = p_tzcf_all_yrs_2006_2007_2023, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

gg_tzcf_all_yrs_2006_2007_2023 <- ggplotly(p_tzcf_all_yrs_2006_2007_2023, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))

```


```{r 04-plot-tzcf-2003-2011-2023-yrs}
p_tzcf_all_yrs <-  ggplot() +
    # Create additional contour layers (replace with your data and aesthetics)
       # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year != 2023), col="Individual Years (1997-2022)") +
    
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2003), col="2003") +
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2004), col="2004") +
        add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2005), col="2005") +
     add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2006), col="2006") +
     add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2007), col="2007") +
    # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2008), col="2008") +
    # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2009), col="2009") +
    add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2010), col="2010") +
    add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year == 2011), col="2011") +
    
        add_tzcf_contour_2023(col="2023") + 
    
        add_baseplot_theme() +
    
    scale_y_continuous(breaks=seq(30, 45, 2)) +


        scale_color_manual(
        name = "18°C Isotherm (165°W-140°W)\n ",
        values = c("#4CC9F0","#4895EF","#4361EE", "#3F37C9", "#3A0CA3","#560BAD", "#7209B7", "#F72585")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("Individual Years (1997-2022)", "2015","2023")  # Specify labels
        )


gg_tzcf_all_yrs <- ggplotly(p_tzcf_all_yrs, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))


```

```{r 04-plot-tzcf-enso-yrs}

p_tzcf_enso_yrs <- 
ggplot() +
    # Create additional contour layers (replace with your data and aesthetics)
       # add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year != 2023), col="Individual Years (1997-2022)") +
    
        add_tzcf_indiv_contours(grouped_by_year_tzcf_enso_all %>% filter(enso == 'el_nino'), col="El Nino") +
        add_tzcf_indiv_contours(grouped_by_year_tzcf_enso_all %>% filter(enso == 'la_nina'), col="La Nina") +
        add_tzcf_indiv_contours(grouped_by_year_tzcf_enso_all %>% filter(enso == 'neutral'), col="Neutral") +
         add_tzcf_indiv_contours(grouped_by_year_tzcf_enso_all %>% filter(enso == 'mhw'), col="MHW") +
    
        # add_tzcf_contour_2023(col="2023") + 
    #     geom_ribbon(data = grouped_by_year_tzcf_enso_all %>% filter(enso == 'el_nino'), 
    #                 aes(x = x, ymin=z_min, ymax = z_max), linetype = 0, alpha = 0.5)+
    #     # geom_ribbon(aes(ymin=xmin, ymax = xmax, fill = type), linetype = 0, alpha =0.5)+
         # add_tzcf_indiv_contours(grouped_by_year_tzcf_enso_all %>% filter(enso == 'mhw'), col="MHW") +
# 
#     geom_contour(data=grouped_by_year_tzcf_enso_all %>% filter(enso == 'mhw'),
#         aes(x=month, y=y, z = z_min, text = paste("Year: ", year), colour = "MHW"),
#         linewidth = 1.25, breaks = c(18)) +
    
        add_baseplot_theme() +
    
    scale_y_continuous(breaks=seq(30, 45, 2)) +


        scale_color_manual(
        name = "18°C Isotherm\nAvg Latitude\n(165°W-140°W)\n ",
        breaks = c("El Nino","La Nina", "Neutral", "MHW"),
        values = c("El Nino" = "tomato",
                   "La Nina" = "#4895EF",
                   "Neutral" = "azure3", 
                   "MHW" = "orange")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("Individual Years (1997-2022)", "2015","2023")  # Specify labels
        )

# save
ggsave(filename = 'plot_tzcf_enso_yrs.png', plot = p_tzcf_enso_yrs + 
               theme(plot.caption = element_text(hjust = 0)) +
                labs(caption="Average location of 18°C SST isotherm by month (1985 - 2023)"), 
       # height=4,
       width=10,
       bg='white',
       path = '../figs/'
       )

```



```{r 04-plot-tzcf-pre-2010-post-yrs}
p_tzcf_all_yrs_split <-  ggplot() +
    # Create additional contour layers (replace with your data and aesthetics)
       add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year < 2010), col="Individual Years (1997-2009)") +
           add_tzcf_indiv_contours(grouped_by_year_month %>% filter(year >= 2010), col="Individual Years (2010-2022)") +

        add_tzcf_contour_2023(col="2023") + 
    
        add_baseplot_theme() +
    
    scale_y_continuous(breaks=seq(30, 45, 2)) +

        scale_color_manual(
        name = "18°C Isotherm (165°W-140°W)\n ",
        values = c( "#952C23", "gray80","#F8C08C" )#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("Individual Years (1997-2022)", "2015","2023")  # Specify labels
        ) 


gg_tzcf_all_yrs_split <- ggplotly(p_tzcf_all_yrs_split, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))


```


```{r 04-calc-mlat-avgs-cc}
mlats_historic <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
        # filter(lon >= -179 & lon <= -130) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat))

mlats_historic_2005_2006 <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(year >= 2005 & year <= 2006) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))

mlats_historic_2006_2007 <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(year >= 2006 & year <= 2007) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat))
    
mlats_cohort1 <- daily_avg_data_cohort_1 %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat)) 
```


```{r 04-cross-check-mlats-historic-with-02-weekly-lat-avgs}
# # fyi, will need to load 02 rmd RData
# weekly_hist_df_avg_lats <- 
# df_165W %>%
#     filter(group == 'historic') %>%
#     mutate(
#     week_idx = week(date)) %>%
#     mutate(week_date = getWeekDate(week_idx)) %>%
#     group_by(group, week_date) %>%
#     summarize(
#         avg = mean(lat, na.rm=TRUE))
# 
# 
# monthly_hist_df_avg_lats <- weekly_hist_df_avg_lats %>%
#     mutate(year = lubridate::year(week_date),
#            month = lubridate::month(week_date)) %>%
#     group_by(month) %>%
#        summarize(
#         avgs = mean(avg, na.rm=TRUE))

```


```{r 04-calc-n-mlats-historic-per-month}

n_mlats_historic_per_month <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
        # filter(lon >= -179 & lon <= -130) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(n_distinct(id), .groups = "drop")

n_mlats_historic_2005_2006_per_month <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(year >= 2005 & year <= 2006) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(n_distinct(id), .groups = "drop")


n_mlats_cohort_1_per_month <- daily_avg_data_cohort_1 %>%
    filter(lon >= -165 & lon <= -140) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(n_distinct(id), .groups = "drop")
```


```{r 04-gg-seasonal-tzcf-yrs-only}
p_tzcf_hist_yrs_cohort_yr <-  ggplot() +
 
        add_tzcf_contour(grouped_by_year_tzcf_1997_2012, col="Average (1997-2012)") +
    
        add_tzcf_contour(grouped_by_year_tzcf_2005_2006, col="Average (2005-2006)") +
    
        add_tzcf_contour_2023(col='2023') +
        
        add_baseplot_theme() +
    
        scale_y_continuous(breaks=seq(30, 45, 2)) +
    
        scale_color_manual(
        name = "18°C Isotherm (165°W-140°W)\n ",
        values = c("gray80","#E4938B", "#952C23")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("1997-2022", "2023")  # Specify labels
        )


gg_tzcf_hist_yrs_cohort_yr <- ggplotly(p_tzcf_hist_yrs_cohort_yr, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))

```


```{r 04-gg_tzcf_hist_cohort}
p_tzcf_hist_cohort <- 
    ggplot() +
# p_tzcf_hist_yrs_cohort_yr + 
    
        # add_tzcf_contour(grouped_by_year_tzcf_2005_2006, col="18°C Isotherm (2005-2006)") +
        add_tzcf_contour_2023(col='18°C Isotherm (2023)') +
        add_cohort1_mlats(col="Cohort 1 Tracks (2023)") +
    
        add_tzcf_contour(grouped_by_year_tzcf_1997_2012, col="18°C Isotherm (1997-2012)") +
        
        add_historic_mlats(col="Historic Tracks (1997-2012)") + 
        add_historic_2005_2006_mlats(col="Historic Tracks (2005-2006)") +

        add_baseplot_theme() +
    
        scale_y_continuous(breaks=seq(30, 45, 2)) +
    
    scale_color_manual(
        name = "Avg Monthly Lat (165°W-140°W)\n ",
        breaks = c("18°C Isotherm (2023)",
                   "18°C Isotherm (1997-2012)", 
                   
                   "Historic Tracks (1997-2012)", 
                   "Historic Tracks (2005-2006)",
                   "Cohort 1 Tracks (2023)"                  
                    ),

                   
        values = c(
                   "18°C Isotherm (2023)"="gray30",
                   "18°C Isotherm (1997-2012)"="gray80",
                   
                   "Historic Tracks (1997-2012)"="#6D4C94",
                   "Historic Tracks (2005-2006)"="#E05780",
                   "Cohort 1 Tracks (2023)"="#66CCC0"
                   )
    #,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("2023: 18°C Isotherm", "1997-2022: 18°C Isotherm","Historic Tracks", "Cohort 1 Tracks")  # Specify labels
        )

 gg_tzcf_hist_cohort <- ggplotly(p_tzcf_hist_cohort, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))
    
```


```{r 04-gg_tzcf_hist_cohort-sept-lag}
p_tzcf_hist_cohort_2006_2007_2023 <- 
    ggplot() +
# p_tzcf_hist_yrs_cohort_yr + 
    
        # add_tzcf_contour(grouped_by_year_tzcf_2005_2006, col="18°C Isotherm (2005-2006)") +
        add_tzcf_contour_2023(col='18°C Isotherm (2023)') +
        add_cohort1_mlats(col="Cohort 1 Tracks (2023)") +
    
        add_tzcf_contour(grouped_by_year_tzcf_1997_2012, col="18°C Isotherm (1997-2012)") +
        
        add_historic_mlats(col="Historic Tracks (1997-2012)") + 
        add_historic_2005_2006_mlats(col="Historic Tracks (2006-2007)") +

        add_baseplot_theme() +
    
        scale_y_continuous(breaks=seq(30, 45, 2)) +
    
    scale_color_manual(
        name = "Avg Monthly Lat (165°W-140°W)\n ",
        breaks = c("18°C Isotherm (2023)",
                   "18°C Isotherm (1997-2012)", 
                   
                   "Historic Tracks (1997-2012)", 
                   "Historic Tracks (2006-2007)",
                   "Cohort 1 Tracks (2023)"                  
                    ),

                   
        values = c(
                   "18°C Isotherm (2023)"="gray30",
                   "18°C Isotherm (1997-2012)"="gray80",
                   
                   "Historic Tracks (1997-2012)"="#6D4C94",
                   "Historic Tracks (2005-2006)"="#E05780",
                   "Cohort 1 Tracks (2023)"="#66CCC0"
                   )
    #,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("2023: 18°C Isotherm", "1997-2022: 18°C Isotherm","Historic Tracks", "Cohort 1 Tracks")  # Specify labels
        )

 gg_tzcf_hist_cohort_2006_2007_2023 <- ggplotly(p_tzcf_hist_cohort_2006_2007_2023, tooltip = c("text")) %>% layout(xaxis = list(title = "Month"), yaxis = list(title = "Latitude (°N)"))
    
```

<!-- ```{r 04-plot-ssta-cohort-1-with-contour} -->
<!-- p_ssta_mean_tcms_w_contour <- p_ssta_mean_tcms +  -->
<!--     add_contour_lines(df=ssta_ras_mean_df, contours=c(1.5)) +  -->
<!--     resize_cbar(14) -->

<!-- ``` -->

```{r 04-interim-rdata-save}

## INTERIM SAVE 
f <- "../data/interim/04_northern_lats_EDA_interim_data.RData"
if (file.exists(f)){
    invisible()
} else{
    save.image(file = "../data/interim/04_northern_lats_EDA_interim_data.RData")
}

```
