==> rmd/05_northern_lats_EDA_gam.Rmd <==
---
title: "5 | Fit GAM"
subtitle: "Project STRETCH: Northern Latitudes - Exploratory Analyses"
author: "Dana K Briscoe"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: none
    df_print: paged
    highlight: tango
    number_sections: no
    theme: flatly
    toc: no
    toc_float: true
    toc_collapsed: true
  pdf_document:
    toc: yes
    toc_depth: '2'
---

```{=html}
<style>
.main-container {
    max-width: 1180px;
    margin-left: 10;
    margin-right: 10;
}
</style>
```

```{r 05-setup, include=FALSE}
knitr::opts_chunk$set(
    echo = FALSE,
    eval = TRUE,
    message = FALSE,
    warning = FALSE,
    dpi = 300,
    fig.align = "center"
    )

# See Options: knitr::opts_chunk$get()
```

```{r 05-load-libraries, message=FALSE}
# 05
library(tidyverse)
library(data.table)
library(mgcv)
library(ggalt)
library(scales)
library(plotly)
library("gratia")
library(glue)
```

```{r 05-source-helper-funcs}
source('../code/00_northern_lats_helper_functions.R')
```

```{r 05-load-cohort-1}
source('../code/01_prep_turtle_data.R')

```

```{r 05-rename-dfs-for-consistency}
raw_df <- raw_data_cohort_1_w_names
daily_df <- daily_avg_data_cohort_1

# make historic df consistent with cohort df(s)
hist_df <-  historic_tags %>% wrangleHistDF()

```

```{r 05-subset-hist-df-septs}
hist_df_subset <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    mutate(mon_yr = str_c(month, "-",year))

# length(unique(hist_df_subset$id))

hist_df_subset_septs <- hist_df_subset %>% filter(month == 9)
# length(unique(hist_df_subset_septs$id))

```


<!-- ```{r} -->

<!-- test =  -->
<!-- hist_df %>% -->
<!--     filter(lon >= -165 & lon <= -140) %>% -->
<!--     group_by(id) %>% -->
<!--     group_modify(~binFreqTable(.$lon, seq(-165,-140, 5))) %>% -->
<!--     # group_by(range) -->
<!--     ungroup() -->

<!-- test |> summarise(id |> unique() |> length(), .by = range) -->
<!-- ``` -->



```{r 05-ntags-165W-140W-per-month}
ntags_hist_tags_165W_140W_per_month <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    group_by(month) %>%
    summarize(count=n_distinct(id)) %>%
    # mutate(month = month.abb[month]) %>%
    spread(month, count)

colnames(ntags_hist_tags_165W_140W_per_month) <- month.abb
```


```{r 05-ntags-165W-140W-per-year}
ntags_hist_tags_165W_140W_per_year <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    # filter(month == 9) %>%
    group_by(year) %>%
    summarize(count=n_distinct(id)) %>%
    spread(year, count)

```

```{r 05-ntags-165W-140W-per-year-month-sept-only}
ntags_hist_tags_165W_140W_per_year_month_sept <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 3) %>%
    group_by(year) %>%
    summarize(count=n_distinct(id)) %>%
    spread(year, count)

```


```{r 05-ntags-165W-140W-per-year-month}
ntags_hist_tags_165W_140W_per_year_month <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    group_by(year, month) %>%
    summarize(count=n_distinct(id)) %>%
    spread(month, count)

colnames(ntags_hist_tags_165W_140W_per_year_month) <- c("Year", month.abb)
```


```{r 05-calc-max-lat-avgs-cc-sept}
max_lats_historic <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 9) %>%
    # group_by(year) %>%
    # summarize(count=n_distinct(id)) %>%
    group_by(id) %>%
    slice(which.max(lat)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = as.character(id)) %>%
    ungroup() %>%
    as_tibble()

max_lats_cohort1 <- daily_avg_data_cohort_1 %>%
    filter(month == 9) %>%
    group_by(id) %>%
    slice(which.max(lat)) %>%
    mutate(lon = round(lon, 3),
           lat = round(lat, 3)) %>%
    mutate(group = as.character(id)) %>%
    ungroup() %>%
    as_tibble()

max_lats_all <- rbind(max_lats_historic, max_lats_cohort1)
```

```{r 04-repeat-calc-mlat-avgs-cc}
mlats_historic <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
        # filter(lon >= -179 & lon <= -130) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat))

mlats_historic_2005_2006 <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(year >= 2005 & year <= 2006) %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat))
    
mlats_cohort1 <- daily_avg_data_cohort_1 %>%
    dplyr::select(-c(date)) %>%
    group_by(month) %>%
    summarise(avgs = mean(lat, na.rm=T))
    # summarise(avgs = median(lat)) 


## Now do the same but calc monthly lat avgs by month & year
mlats_historic_mon_year <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    dplyr::select(-c(date)) %>%
    group_by(month, year) %>%
    summarise(avg_lat = mean(lat, na.rm=T), .groups = 'drop') %>%
    mutate(group = 'historic') 
    
mlats_cohort1_mon_year <- daily_avg_data_cohort_1 %>%
    filter(lon >= -165 & lon <= -140) %>%
    dplyr::select(-c(date)) %>%
    group_by(month, year) %>%
    summarise(avg_lat = mean(lat, na.rm=T), .groups = 'drop') %>%
    mutate(group = 'cohort1')

mlats_all_mon_year <- rbind(mlats_historic_mon_year, mlats_cohort1_mon_year)
    
```


```{r 05-gg-graph-hist-id-yrmon}

# slice into 2 plots, since there are 69 ids

p_graph_hist_id_by_yrmon <- ggplot() +
  geom_line(data = hist_df_subset, aes(x=date, y=id, group=id)) +
    geom_line(data = hist_df_subset %>% filter(month == 9), aes(x=date, y=id, group=id, color='September'), lwd=2) +
    scale_color_manual(values=c("#952C23")) +
    guides(color=guide_legend(title="Tracks\n165W-140W\nLongitude")) +
    scale_x_date(breaks = "3 month", labels=date_format("%b-%Y")) +
   theme(axis.text.x = element_text(angle = 45, vjust = 1.0, hjust=1,size=6),
         axis.text.y = element_text(size=6))


# save
ggsave(filename = 'ntags_historic_data_165W_140W_Sept_months.png', plot = p_graph_hist_id_by_yrmon + labs(subtitle=""), 
       # height=4,
       width=13,
       bg='white',
       path = '../figs/'
       )

gg_graph_hist_id_by_yrmon<- ggplotly(p_graph_hist_id_by_yrmon, tooltip = c("text")) %>% layout(xaxis = list(title = "Date"), yaxis = list(title = "Historic Turtle Data ID Number"))

```


```{r 05-max-lat-yr-avg-meds}
max_lat_yr_w2023 <- max_lats_all %>%
    filter(lon <= -140) %>%
    # filter(year != 2023) %>%
    group_by(year) %>% 
    summarise(avg_lon = mean(lon, na.rm=T),
              avg_lat = mean(lat, na.rm=T),
              med_lon = median(lon, na.rm=T),
              med_lat = median(lat, na.rm=T))

max_lat_yr_wo_2023 <- max_lats_all %>%
    filter(lon <= -140) %>%
    filter(year != 2023) %>%
    group_by(year) %>% 
    summarise(avg_lon = mean(lon, na.rm=T),
              avg_lat = mean(lat, na.rm=T),
              med_lon = median(lon, na.rm=T),
              med_lat = median(lat, na.rm=T))


```


```{r 05-p-histogram-hist-df-septs-avg-medians}
hist_all_stats <- hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 9) %>%
    summarise(mean = round(mean(lat, na.rm=T),2),
              med = round(median(lat, na.rm=T),2)) %>%
    gather(., 'stat', value)

p_hist_df_avg_vs_med <- 
hist_df %>%
    filter(lon >= -165 & lon <= -140) %>%
    filter(month == 9) %>%
ggplot(aes(lat)) +    
geom_histogram(binwidth = 0.5, fill = "steelblue", colour = "navy", alpha = 0.6) + 
    geom_vline(data = hist_all_stats, aes(xintercept = value, color = stat),  linewidth=0.9) +
    scale_color_manual(values = c("red", "orange"), labels = c("Mean", "Median"), name = "Statistic")  +
    labs(x = 'Latitude (°N)', y= "Count") +
    theme_bw()

```



<!-- ```{r try-gams} -->


<!-- ## RUN MODELS -->
<!-- # mean -->
<!-- gam_model_w2023_mean <- gam(avg_lat ~ s(avg_lon, k = 3) + s(year, k = 3), data = max_lat_yr_w2023) -->
<!-- gam_model_wo_2023_mean <- gam(avg_lat ~ s(avg_lon, k = 3) + s(year, k = 3), data = max_lat_yr_wo_2023) -->

<!-- # median -->
<!-- gam_model_w2023 <- gam(med_lat ~ s(med_lon, k = 3) + s(year, k = 3), data = max_lat_yr_w2023) -->
<!-- gam_model_wo_2023 <- gam(med_lat ~ s(med_lon, k = 3) + s(year, k = 3), data = max_lat_yr_wo_2023) -->


<!-- # # # # gam_model <- gam_model_w2023 -->
<!-- # # # # Summary of the GAM model -->
<!-- # summary(gam_model_w2023_mean) -->
<!-- # summary(gam_model_wo_2023_mean) -->
<!-- #  -->
<!-- # summary(gam_model_w2023) -->
<!-- # summary(gam_model_wo_2023) -->
<!-- # # #  -->
<!-- # # # # Display partial plots for "lon" and "year" -->
<!-- # # # par(mfrow=c(2, 2))  # Set up a grid for two plots side by side -->
<!-- # # # plot(gam_model_w2023, select = 1, shade = TRUE, col = "blue", main = "Partial Plot for Lon (incl 2023)") -->
<!-- # # # plot(gam_model_w2023, select = 2, shade = TRUE, col = "red", main = "Partial Plot for Year (incl 2023)") -->
<!-- # # #  -->
<!-- # # # plot(gam_model_wo_2023, select = 1, shade = TRUE, col = "blue", main = "Partial Plot for Lon (without 2023)") -->
<!-- # # # plot(gam_model_wo_2023, select = 2, shade = TRUE, col = "red", main = "Partial Plot for Year (without 2023)") -->
<!-- # #  -->
<!-- # ## MEAN -->
<!-- # draw(gam_model_w2023_mean, residuals = TRUE) -->
<!-- # draw(gam_model_wo_2023_mean, residuals = TRUE) -->
<!-- #  -->
<!-- # # par(mfrow=c(1, 2))  # Set up a grid for two plots side by side -->
<!-- # draw(gam_model_w2023, residuals = TRUE) -->
<!-- # draw(gam_model_wo_2023, residuals = TRUE) -->

<!-- ``` -->

<!-- ```{r} -->

<!-- # Create the GAM model -->
<!-- gam_model <-  gam(avg_lat ~ s(avg_lon, k = 3) + s(year, k = 3), data = max_lat_yr_w2023) -->

<!-- # Initialize a vector to store cross-validation results -->
<!-- cv_results <- numeric(nrow(max_lat_yr_w2023)) -->

<!-- # Perform leave-one-out cross-validation -->
<!-- for (i in 1:nrow(max_lat_yr_w2023)) { -->
<!--   # Exclude the i-th observation -->
<!--   data_subset <- max_lat_yr_w2023[-i, ] -->

<!--   # Fit the model on the subset -->
<!--   model_subset <-  gam(avg_lat ~ s(avg_lon, k = 4) + s(year, k = 4), data = max_lat_yr_w2023) -->

<!--   # Predict the excluded observation -->
<!--   prediction <- predict(model_subset, newdata = max_lat_yr_w2023[i, ]) -->

<!--   # Calculate and store the squared prediction error -->
<!--   cv_results[i] <- (max_lat_yr_w2023[i, "avg_lat"] - prediction)^2 -->
<!-- } -->

<!-- # Calculate the mean squared prediction error (MSE) for LOOCV -->
<!-- cv_results_df <- do.call(rbind, cv_results) -->

<!-- mse_cv <- mean(cv_results_df, na.rm=TRUE) -->

<!-- # # View the LOOCV MSE -->
<!-- # print(mse_cv) -->

<!-- ``` -->

```{r 05-load-grouped-by-year-month-rds}

# fyi, need to load this df from '04 rmd'
grouped_by_year_month <- readRDS("../data/interim/04_grouped_by_year_month_Jan1981_Oct2023.rds")

```

```{r 05-get-sum-stats-lat-year-sept-mar}

# first subset to only keep Mar & Sept; then group by month and get summary stats
grouped_by_year_month_subset <- grouped_by_year_month %>%
    filter(month == 3 | month == 9) 
    # split(grouped_by_year_month_subset$month)

# get min, mean, max lats of 18C per year
min_18deg_lat_by_year <- grouped_by_year_month_subset %>%
    group_by(year, month) %>%
    # filter(month == 9) %>%
    filter(z >= 18 & z < 18.1) %>%
    # filter(z == 18) %>%  # fyi, this returns no values, so go up to 18.05 or 18.1
    slice(which.min(y)) %>%
    mutate(min_18degC_lat = y)

mean_18deg_lat_by_year <- grouped_by_year_month %>%
    group_by(year, x, month, day, month_abb) %>%
        # filter(month == 9) %>%
    filter(z >= 18 & z < 18.1) %>%
    summarise(y = mean(y, na.rm=T),
              z = mean(z, na.rm=T)) %>%
    relocate(year, .after=x) %>%
    relocate(y, .before=x) %>%
    relocate(z, .after=x) %>%
    mutate(mean_18degC_lat = y)
    
max_18deg_lat_by_year <- grouped_by_year_month %>%
    group_by(year, month) %>%
    # filter(month == 9) %>%
    filter(z >= 18 & z < 18.1) %>%
    slice(which.max(y)) %>%
    mutate(max_18degC_lat = y)
    
 
# Sept 19.5
mean_195deg_lat_by_year_sept <- grouped_by_year_month %>%
    group_by(year, x, month, day, month_abb) %>%
        filter(month == 9) %>%
    filter(z >= 19.5 & z < 19.6) %>%
    summarise(y = mean(y, na.rm=T),
              z = mean(z, na.rm=T)) %>%
    relocate(year, .after=x) %>%
    relocate(y, .before=x) %>%
    relocate(z, .after=x) %>%
    mutate(mean_degC_lat = y)

# Mar 16
mean_16deg_lat_by_year_mar <- grouped_by_year_month %>%
    group_by(year, x, month, day, month_abb) %>%
        filter(month == 3) %>%
    filter(z >= 16.1 & z < 16.2) %>%
    summarise(y = mean(y, na.rm=T),
              z = mean(z, na.rm=T)) %>%
    relocate(year, .after=x) %>%
    relocate(y, .before=x) %>%
    relocate(z, .after=x) %>%
    mutate(mean_degC_lat = y)
   
```

```{r 05-plot-sum-stats-lat-year-sept-mar}
df_sum_stats_lat_year_list <- list(min = min_18deg_lat_by_year, mean = mean_18deg_lat_by_year, max = max_18deg_lat_by_year)

plot_sum_stats_lat_year_by_month <- function(df_list, mon = 9, pt_col="steelblue"){
    p <- ggplot() +
    # geom_point(data = df_list[['min']] %>% filter(month == mon), aes(x=year, y=y, col='Min Lat of 18°C Isotherm',
                      # text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Min")), size=4) +
    geom_point(data = df_list[['mean']] %>% filter(month == mon), aes(x=year, y=y, col='Mean Lat of 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    # geom_point(data = df_list[['max']] %>% filter(month == mon), aes(x=year, y=y, col='Max Lat of 18°C Isotherm',
    #                   text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Max")), size=4) +
       scale_color_manual(
        name = glue("Latitude of {month.abb[mon]} 18°C Isotherm \n(165°W-140°W) by Year\n\n"),
        breaks = c(
                    # 'Max Lat of 18°C Isotherm', 
                   'Mean Lat of 18°C Isotherm'#, 
                   # 'Min Lat of 18°C Isotherm'
                   ),
        values = c(
            # 'Max Lat of 18°C Isotherm'="#952C23",
            # 'Mean Lat of 18°C Isotherm'="#E4938B", 
            'Mean Lat of 18°C Isotherm'= pt_col#, 
            # 'Min Lat of 18°C Isotherm'="steelblue"
                )#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("1997-2022", "2023")  # Specify labels
        ) + 
    geom_line(data = df_list[['mean']] %>% filter(month == mon), aes(x=year, y=y), col='azure4') +
        theme(legend.position = "none") +
    
    scale_x_continuous(breaks=seq(1985,2023, 5)) +
    scale_y_continuous(breaks=seq(29,46,0.5)) +
    labs(x = "Year", y = "Latitude (°N")

    return(p)    
}


```

```{r 05-gg-plot-sum-stats-lat-year-mar}

p_sum_stats_lat_year_mar <- plot_sum_stats_lat_year_by_month(df_list = df_sum_stats_lat_year_list, mon = 3)

gg_sum_stats_lat_year_mar <- ggplotly(p_sum_stats_lat_year_mar, tooltip = c("text")) %>% layout(xaxis = list(title = "Year"), yaxis = list(title = "Latitude (°N)"))

```


```{r 05-gg-plot-sum-stats-lat-year-sept}
p_sum_stats_lat_year_sept <- plot_sum_stats_lat_year_by_month(df_list = df_sum_stats_lat_year_list, mon = 9, pt_col="#952C23")


gg_sum_stats_lat_year_sept <- ggplotly(p_sum_stats_lat_year_sept, tooltip = c("text")) %>% layout(xaxis = list(title = "Year"), yaxis = list(title = "Latitude (°N)"))

```


```{r 05-gg-plot-sum-stats-avg-lats-18C-cc-year-mar-sept}

df_list = df_sum_stats_lat_year_list

p <- ggplot() +

    geom_point(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y, col='March 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    geom_point(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y, col='Sept 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    geom_point(data = mlats_all_mon_year %>% filter(month == 3), aes(x=year, y=avg_lat, col='March Turtles',
                      text = str_c("Year: ", year, "\nLat: ", avg_lat, "°N", "\nStat: Mean")), size=4) + 
        geom_point(data = mlats_all_mon_year %>% filter(month == 9), aes(x=year, y=avg_lat, col='Sept Turtles',
                      text = str_c("Year: ", year, "\nLat: ", avg_lat, "°N", "\nStat: Mean")), size=4) + 
    
       scale_color_manual(
        name = glue("Average Latitude \n(165°W-140°W) \nby Year\n\n"),
        breaks = c('March 18°C Isotherm','Sept 18°C Isotherm',"March Turtles", "Sept Turtles"),
        values = c('March 18°C Isotherm'="#E4938B",'Sept 18°C Isotherm'="#952C23", "March Turtles"= "steelblue1", "Sept Turtles"="steelblue4")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("1997-2022", "2023")  # Specify labels
        ) + 
    geom_line(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y), col='#E4938B') +
    geom_line(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y), col='#952C23') +
    geom_line(data = mlats_all_mon_year %>% filter(month == 3), aes(x=year, y=avg_lat), col='steelblue1') +
    geom_line(data = mlats_all_mon_year %>% filter(month == 9 & year <=2013), aes(x=year, y=avg_lat), col='steelblue4') +
    
    scale_x_continuous(breaks=seq(1997,2023, 1)) +
    scale_y_continuous(breaks=seq(29,46,1)) +
    labs(x = "Year", y = "Latitude (°N") + 
    theme(legend.position="top", legend.title = element_text(hjust=0.5)) +
    guides(fill=guide_legend(nrow=2,byrow=TRUE))

### Save plot ----
ggsave(filename = 'plot_mar_sept_avg_lats_turtles_18C.png', plot = p, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )
```

```{r 05-gg-plot-sum-stats-avg-lats-18C-no-cc-year-mar-sept}

df_list = df_sum_stats_lat_year_list

p <- ggplot() +

    geom_point(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y, col='March 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    geom_point(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y, col='Sept 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    # geom_point(data = mlats_all_mon_year %>% filter(month == 3), aes(x=year, y=avg_lat, col='March Turtles',
    #                   text = str_c("Year: ", year, "\nLat: ", avg_lat, "°N", "\nStat: Mean")), size=4) + 
    #     geom_point(data = mlats_all_mon_year %>% filter(month == 9), aes(x=year, y=avg_lat, col='Sept Turtles',
    #                   text = str_c("Year: ", year, "\nLat: ", avg_lat, "°N", "\nStat: Mean")), size=4) + 
    # 
       scale_color_manual(
        name = glue("Average Latitude \n(165°W-140°W) \nby Year\n\n"),
        breaks = c('March 18°C Isotherm','Sept 18°C Isotherm'),
        values = c('March 18°C Isotherm'="#F48C06",'Sept 18°C Isotherm'="#9D0208")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("1997-2022", "2023")  # Specify labels
        ) + 
    geom_line(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y), col='#F48C06') +
    geom_line(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y), col='#9D0208') +

    
    scale_x_continuous(breaks=seq(1985,2023, 5)) +
    scale_y_continuous(breaks=seq(29,46,1)) +
    labs(x = "Year", y = "Latitude (°N") + 
    theme(legend.position="top", legend.title = element_text(hjust=0.5)) +
    guides(fill=guide_legend(nrow=2,byrow=TRUE))

### Save plot ----
ggsave(filename = 'plot_mar_sept_avg_lats_no_turtles_18C.png', plot = p, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )
```

```{r 05-set-enso-conds}

enso_yr <- data.frame(year = seq(1985,2023,1)) %>%
    mutate(enso = case_when((year == 1986 | year == 1987 | year == 1991 | year == 1994 | year == 1997 | 
                                 year == 2002 | year == 2004 | year == 2006 | year == 2009 | year == 2015 | year == 2018 | year == 2023) ~ "el_nino",
                            
                            (year == 1988 | year == 1995 | year == 1998 | year == 1999 | 
                                 year == 2000 | 
                                 # year == 2005 | 
                                 year == 2007 | 
                                 # year == 2008 | 
                                 year == 2010 | year == 2011 | year == 2016 | year == 2017 | 
                                 year == 2020 | year == 2021 | year == 2022) ~ "la_nina",
                            
                            (year == 2014 | year == 2015 | year == 2019 | year == 2020) ~ "mhw",    
                            TRUE ~ "neutral"))



sept_dte <- getZ(sept_sst_ras_stack)
sept_y <- format(sept_dte, "%Y")

```

```{r 05-gg-plot-sum-stats-avg-lats-18C-no-cc-year-mar-sept-with-enso-conds}

df_list = df_sum_stats_lat_year_list

p_enso <- ggplot() +

    geom_point(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y, col='March 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 
    geom_point(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y, col='Sept 18°C Isotherm',
                      text = str_c("Year: ", year, "\nLat: ", y, "°N", "\nStat: Mean")), size=4) + 

       scale_color_manual(
        name = glue("Average Latitude\nof 18°C isotherm\n\n"),
        breaks = c('March 18°C Isotherm','Sept 18°C Isotherm'),
        values = c('March 18°C Isotherm'="#F48C06",'Sept 18°C Isotherm'="#9D0208")#,  # Specify colors - don't know why this is backwards from labels....
        # labels = c("1997-2022", "2023")  # Specify labels
        ) + 
    geom_line(data = df_list[['mean']] %>% filter(month == 3), aes(x=year, y=y), col='#F48C06') +
    geom_line(data = df_list[['mean']] %>% filter(month == 9), aes(x=year, y=y), col='#9D0208') +

    # el yrs
    geom_rect(data=enso_yr %>% filter(enso == 'el_nino'), aes(xmin=year-.5, xmax=year+.5, ymin=29, ymax=45, 
              fill = "El Nino"), color = NA, alpha = .3) +
    # la yrs
    geom_rect(data=enso_yr %>% filter(enso == 'la_nina'), aes(xmin=year-.5, xmax=year+.5, ymin=29, ymax=45, 
              fill = "La Nina"), color = NA, alpha = .3) +
    
     # mhw yrs
    geom_rect(data=enso_yr %>% filter(enso == 'mhw'), aes(xmin=year-.5, xmax=year+.5, ymin=29, ymax=45, 
              fill = "MHW"), color = NA, alpha = .3) +
    
    scale_fill_manual('ENSO/MWH\nYear',
                      values = c('red', 'dodgerblue', 'orange'),  
                      guide = guide_legend(override.aes = list(alpha = 1))) +
    
    scale_x_continuous(breaks=seq(1985,2023, 2)) +
    scale_y_continuous(breaks=seq(29,46,1)) +
    labs(x = "Year", y = "Average Latitude (°N)") + 
    theme_minimal() +
    theme(legend.position="top", legend.title = element_text(hjust=0.5)) +
    guides(fill=guide_legend(nrow=2,byrow=TRUE))

### Save plot ----
ggsave(filename = 'plot_mar_sept_avg_lats_no_turtles_18C_enso_conds.png', plot = p_enso, 
       # height=9,
       width=13, bg='white',
       path = '../figs/'
       )

```



```{r 05-fit-gam-18deg-lat-by-year- sept}

# df = min_18deg_lat_by_year %>% filter(month == 9)
df = mean_18deg_lat_by_year %>% filter(month == 3)

# Initialize a vector to store cross-validation results
cv_results <- numeric(nrow(df))

model_plots <- list()
# Perform leave-one-out cross-validation
for (i in 1:nrow(df)) {
  # Exclude the i-th observation
  data_subset <- df[-i, ]

  # Fit the model on the subset
  model_subset <-  gam(y ~ s(year, k = 10, fx=T), data = data_subset)

  # Predict the excluded observation
  prediction <- predict(model_subset, newdata = df[i, ])

  # Calculate and store the squared prediction error
  cv_results[i] <- (df[i, "y"] - prediction)^2
  
  # draw(model_subset, residuals = TRUE)
  
}

# Calculate the mean squared prediction error (MSE) for LOOCV
cv_results_df <- do.call(rbind, cv_results)

mse_cv <- mean(cv_results_df, na.rm=TRUE)

# summary(model_subset)
 draw(model_subset, residuals = TRUE)


```

```{r}
# Create the GAM model
gam_model <-  gam(y ~ s(year, k = 35), data = df)

summary(gam_model)

test <- draw(gam_model, residuals = TRUE)
test

```

<!-- ```{r} -->
<!-- # Load the required packages -->
<!-- library(mgcv) -->
<!-- library(caret) -->

<!-- # Assuming 'df' is your data frame -->
<!-- # df = min_18deg_lat_by_year %>% filter(month == 9) -->
<!-- df = mean_18deg_lat_by_year %>% filter(month == 9) -->


<!-- # Define the formula for the model -->
<!-- model_formula <- y ~ s(year, k = 30) -->

<!-- # Create a function to fit the GAM model and calculate RMSE -->
<!-- fit_gam_and_rmse <- function(train_indices) { -->
<!--   train_data <- df[train_indices, ] -->
<!--   test_data <- df[-train_indices, ] -->

<!--   gam_model <- gam(model_formula, data = train_data) -->
<!--   predictions <- predict(gam_model, newdata = test_data) -->

<!--   rmse <- sqrt(mean((test_data$y - predictions)^2)) -->

<!--   return(rmse) -->
<!-- } -->

<!-- # Perform k-fold cross-validation (e.g., 5-fold) -->
<!-- k <- 5 -->
<!-- set.seed(123) # Set a random seed for reproducibility -->
<!-- cv_indices <- createFolds(1:nrow(df), k = k, list = TRUE) -->

<!-- cv_rmse <- sapply(cv_indices, function(indices) { -->
<!--   fit_gam_and_rmse(indices) -->
<!-- }) -->

<!-- # Calculate the mean RMSE -->
<!-- mean_rmse <- mean(cv_rmse) -->
<!-- cat("Mean RMSE:", mean_rmse, "\n") -->

<!-- # Assuming 'gam_model_min' is your GAM model -->
<!-- # View a summary of the GAM model -->
<!-- summary(gam_model_min) -->

<!-- draw(gam_model_min, residuals = TRUE) -->
<!-- # plot(gam_model_min, select = 1, scale = 0, rug = TRUE) -->

<!-- ``` -->


```{r kfolds-cv-func}
# Assuming 'df' is your data frame
# df = min_18deg_lat_by_year %>% filter(month == 9)
# df = mean_18deg_lat_by_year %>% filter(month == 3)


# Load the required packages
library(mgcv)
library(boot)
# library(ggplot2)

# Load your dataset or replace this with your data loading method

### K-FOLDS CV ------------------------------------------------------------
run_gam_kfolds <- function(data, kfolds = 5, knts=3, mon = "March"){
        # Define the formula for the GAM model
    # model_formula <- y ~ s(year, k = 20)
    model_formula <- y ~ s(year, k = knts, fx=T)
        
    
    
    # Create an empty vector to store RMSE values
    rmse_values <- numeric()
    
    # Set the number of folds (e.g., 5-fold cross-validation)
    k <- kfolds
    
    # Create a function to calculate the RMSE for the GAM model
    calculate_rmse <- function(train_indices) {
      train_data <- data[train_indices, ]
      test_data <- data[-train_indices, ]
      
      gam_model <- gam(model_formula, data = train_data, method = "ML")
      predictions <- predict(gam_model, newdata = test_data)
      
      rmse <- sqrt(mean((test_data$y - predictions)^2))
      
      # Create a partial dependence plot for 'year' variable
      partial_plot <- draw(gam_model, residuals = TRUE) + labs(title = glue("Partial Plot: {mon}"))
      # title("Partial Dependence Plot")
      
      return(list(rmse = rmse, partial_plot = partial_plot, model_summary = summary(gam_model)))
    }
    
    # Perform k-fold cross-validation
    set.seed(123)  # Set a random seed for reproducibility
    folds <- sample(1:k, nrow(data), replace = TRUE)
    
    results_list <- list()
    
    for (i in 1:k) {
      results <- calculate_rmse(folds != i)
      rmse_values <- c(rmse_values, results$rmse)
      print(results$partial_plot)
      print(results$model_summary)
      
      results_list[[i]] <- list(k=i,results)#, rmse_values, partial_plot=results$partial_plot)
    }
    
    # View the RMSE values for each fold
    cat("RMSE values for each fold:\n")
    print(rmse_values)
    
    # Calculate the mean RMSE
    mean_rmse <- mean(rmse_values)
    cat("Mean RMSE:", mean_rmse, "\n")
    
    return(results_list)
}

```


```{r 05-gam-18degC-lat-year-output-mar}
gam_18degC_lat_year_output_mar <- run_gam_kfolds(data=mean_18deg_lat_by_year %>% filter(month == 3), knts=7, kfolds = 5, mon = "March")

gam_18degC_lat_year_output_mar[[5]][[2]]$partial_plot
# 0.5328260 0.5377308 0.5418552 0.5318179 0.5363884
# Mean RMSE: 0.5361237 
```
```{r}
for(k in 1:5){
    gam_18degC_lat_year_output_mar[[k]][[2]]$partial_plot
}
```



```{r 05-gam-18degC-lat-year-output-sept}
gam_18degC_lat_year_output_sept <- run_gam_kfolds(data=mean_18deg_lat_by_year %>% filter(month == 9), knts=7, kfolds = 5, mon = "September")

gam_18degC_lat_year_output_sept[[4]][[2]]$partial_plot
# 0.9500338 0.9541908 1.0188342 0.9488352 0.9474129
# Mean RMSE: 0.9638614 
```

```{r 05-gam-18degC-lat-year-output-diff-sept-mar}

mean_18deg_lat_by_year_sept_only <- mean_18deg_lat_by_year %>% filter(month == 9)
mean_18deg_lat_by_year_mar_only <- mean_18deg_lat_by_year %>% filter(month == 3)

mean_18deg_lat_by_year_diff_sept_mar <- 
    mean_18deg_lat_by_year %>% filter(month == 9 | month == 3) %>%
    ungroup() %>%
    dplyr::select(c("year", "month_abb", "mean_18degC_lat")) %>%
    # filter(year > 1997) %>% # fyi, there is no march '97, so remove full year
    spread(month_abb,mean_18degC_lat) %>%
    mutate(y = Sep - Mar)

gam_18degC_lat_year_output_diff_sept_mar <- run_gam_kfolds(data=mean_18deg_lat_by_year_diff_sept_mar, knts=5, kfolds = 5, mon = "Sept-Mar")

gam_18degC_lat_year_output_diff_sept_mar[[4]][[2]]$partial_plot
# 0.9786374 1.0676246 1.0400734 0.9769312 0.9957212
# Mean RMSE: 1.011798 

```


```{r 05-plot-mean_18deg_lat_by_year_diff_sept_mar}
library(ggpmisc)

p_mean_18deg_lat_by_year_diff_sept_mar <- 
mean_18deg_lat_by_year_diff_sept_mar %>%
ggplot(aes(x=year, y=y)) +    
    geom_line(fill = "steelblue", colour = "steelblue", alpha = 0.9, lwd =2) +
    geom_point(fill = "steelblue", colour = "steelblue", alpha = 0.6, size=3) + 

    scale_x_continuous(breaks=seq(1985,2023,2)) + 
    labs(x='Year', y = 'Sept - Mar Difference (°Latitude)') +
    
  geom_smooth(method="lm", fit = y ~ x,) 


# save
ggsave(filename = 'draft_fig4x_plot_mean_18deg_lat_by_year_diff_sept_mar_lm.png', plot = p_mean_18deg_lat_by_year_diff_sept_mar + labs(subtitle=""), 
       # height=4,
       width=11,
       bg='white',
       path = '../figs/'
       )

```


```{r kfolds-cv-func-by-ENSO}
# Assuming 'df' is your data frame
# df = min_18deg_lat_by_year %>% filter(month == 9)
# df = mean_18deg_lat_by_year %>% filter(month == 3)


# Load the required packages
library(mgcv)
library(boot)
# library(ggplot2)

# Load your dataset or replace this with your data loading method

### K-FOLDS CV ------------------------------------------------------------
run_gam_kfolds <- function(data, kfolds = 5, knts=3, mon = "March"){
        # Define the formula for the GAM model
    # model_formula <- y ~ s(year, k = 20)
    model_formula <- y ~ s(year, k = knts, fx=T)
        
    
    
    # Create an empty vector to store RMSE values
    rmse_values <- numeric()
    
    # Set the number of folds (e.g., 5-fold cross-validation)
    k <- kfolds
    
    # Create a function to calculate the RMSE for the GAM model
    calculate_rmse <- function(train_indices) {
      train_data <- data[train_indices, ]
      test_data <- data[-train_indices, ]
      
      gam_model <- gam(model_formula, data = train_data, method = "ML")
      predictions <- predict(gam_model, newdata = test_data)
      
      rmse <- sqrt(mean((test_data$y - predictions)^2))
      
      # Create a partial dependence plot for 'year' variable
      partial_plot <- draw(gam_model, residuals = TRUE) + labs(title = glue("Partial Plot: {mon}"))
      # title("Partial Dependence Plot")
      
      return(list(rmse = rmse, partial_plot = partial_plot, model_summary = summary(gam_model)))
    }
    
    # Perform k-fold cross-validation
    set.seed(123)  # Set a random seed for reproducibility
    folds <- sample(1:k, nrow(data), replace = TRUE)
    
    results_list <- list()
    
    for (i in 1:k) {
      results <- calculate_rmse(folds != i)
      rmse_values <- c(rmse_values, results$rmse)
      print(results$partial_plot)
      print(results$model_summary)
      
      results_list[[i]] <- list(k=i,results)#, rmse_values, partial_plot=results$partial_plot)
    }
    
    # View the RMSE values for each fold
    cat("RMSE values for each fold:\n")
    print(rmse_values)
    
    # Calculate the mean RMSE
    mean_rmse <- mean(rmse_values)
    cat("Mean RMSE:", mean_rmse, "\n")
    
    return(results_list)
}

```



<!-- ```{r 05-gam-18degC-lat-year-output-mar-16C} -->
<!-- gam_16degC_lat_year_output_mar <- run_gam_kfolds(data=mean_16deg_lat_by_year_mar %>% filter(month == 3), kfolds = 5, mon = "March") -->

<!-- # RMSE values for each fold: -->
<!-- # [1] 0.5444993 0.5481282 0.5620808 0.5510543 0.5465479 -->
<!-- # Mean RMSE: 0.5504621  -->
<!-- ``` -->

<!-- ```{r 05-gam-18degC-lat-year-output-sept-195C} -->
<!-- gam_195degC_lat_year_output_sept <- run_gam_kfolds(data=mean_195deg_lat_by_year_sept %>% filter(month == 9), kfolds = 5, mon = "September") -->

<!-- gam_18degC_lat_year_output_sept[[4]][[2]]$partial_plot -->
<!-- # [1] 0.8716073 0.8720762 0.8644750 0.8683384 0.8686613 -->
<!-- # Mean RMSE: 0.8690316   -->
<!-- ``` -->

<!-- ```{r 05-gam-195degC-16degC-lat-year-output-diff-sept-mar} -->

<!-- mean_195deg_16deg_lat_by_year_sept_mar <- rbind(mean_195deg_lat_by_year_sept, mean_16deg_lat_by_year_mar) -->

<!-- mean_195deg_16degC_lat_by_year_diff_sept_mar <-  -->
<!--     # mean_18deg_lat_by_year %>% filter(month == 9 | month == 3) %>% -->
<!--     mean_195deg_16deg_lat_by_year_sept_mar %>% -->
<!--     ungroup() %>% -->
<!--     dplyr::select(c("year", "month_abb", "mean_degC_lat")) %>% -->
<!--     filter(year > 1997) %>% # fyi, there is no march '97, so remove full year -->
<!--     spread(month_abb,mean_degC_lat) %>% -->
<!--     mutate(y = Sep - Mar) -->

<!-- gam_19degC_16degC_lat_year_output_diff_sept_mar <- run_gam_kfolds(data=mean_195deg_16degC_lat_by_year_diff_sept_mar, kfolds = 5, mon = "Sept-Mar") -->

<!-- gam_19degC_16degC_lat_year_output_diff_sept_mar[[4]][[2]]$partial_plot -->
<!-- # [1] 1.127050 1.156066 1.180508 1.148367 1.148215 -->
<!-- # Mean RMSE: 1.152041  -->

<!-- ``` -->


<!-- ```{r} -->
<!-- # Split the data into training and testing sets -->
<!-- library(ggplot2) -->
<!-- library(tidyr) -->
<!-- library(caret) -->

<!-- data = mean_195deg_lat_by_year_sept  -->

<!-- set.seed(123)  # for reproducibility -->
<!-- train_index <- createDataPartition(data$mean_degC_lat, p = 0.8, list = FALSE) -->
<!-- train_data <- data[train_index, ] -->
<!-- test_data <- data[-train_index, ] -->

<!-- # Define k-fold cross-validation -->
<!-- folds <- createFolds(train_data$mean_degC_lat, k = 5, list = TRUE, returnTrain = FALSE) -->

<!-- # Build a linear regression model with k-fold cross-validation -->
<!-- lm_model <- train(mean_degC_lat ~ year, data = train_data, method = "lm", trControl = trainControl(method = "cv", index = folds)) -->

<!-- # Make predictions on the training set -->
<!-- train_predictions <- predict(lm_model, newdata = train_data) -->

<!-- # Make predictions on the test set -->
<!-- test_predictions <- predict(lm_model, newdata = test_data) -->

<!-- # Evaluate the model -->
<!-- train_mse <- mean((train_data$mean_degC_lat - train_predictions)^2) -->
<!-- test_mse <- mean((test_data$mean_degC_lat - test_predictions)^2) -->

<!-- cat("Train Mean Squared Error:", train_mse, "\n") -->
<!-- cat("Test Mean Squared Error:", test_mse, "\n") -->

<!-- # Visualize the results -->
<!-- ggplot() + -->
<!--   geom_point(data = train_data, aes(x = year, y = mean_degC_lat), color = 'blue', alpha = 0.5, size = 2) + -->
<!--   geom_line(data = train_data, aes(x = year, y = train_predictions), color = 'red', linetype = 'dashed') + -->
<!--   geom_point(data = test_data, aes(x = year, y = mean_degC_lat), color = 'green', size = 2) + -->
<!--   geom_line(data = test_data, aes(x = year, y = test_predictions), color = 'orange') + -->
<!--   labs(title = "September SST Prediction", -->
<!--        x = "Year", -->
<!--        y = "September SST") -->
<!-- # # Predict SST for the years 2022 and 2023 -->
<!-- # new_data <- data.frame(year = c(2022, 2023)) -->
<!-- # new_predictions <- predict(lm_model, newdata = new_data) -->
<!-- #  -->
<!-- # cat("Predicted SST for 2022:", new_predictions[1], "\n") -->
<!-- # cat("Predicted SST for 2023:", new_predictions[2], "\n") -->

<!-- ``` -->


```{r 05-interim-rdata-save}

## INTERIM SAVE 
f <- "../data/interim/05_northern_lats_EDA_interim_data.RData"
if (file.exists(f)){
    invisible()
} else{
    save.image(file = "../data/interim/05_northern_lats_EDA_interim_data.RData")
}

```
